split huge tests file
make helper libraries separately
