+++ ydb/core/kqp/ut/olap/kqp_olap_ut.cpp
@@ -1,460 +1,32 @@
-#include <ydb/core/kqp/ut/common/kqp_ut_common.h>
-#include <ydb/core/kqp/ut/common/columnshard.h>
-
-#include <ydb/core/sys_view/service/query_history.h>
-#include <ydb/core/tx/columnshard/columnshard_ut_common.h>
-
-#include <contrib/libs/apache/arrow/cpp/src/arrow/api.h>
-#include <contrib/libs/apache/arrow/cpp/src/arrow/ipc/writer.h>
-
-#include <ydb/library/yql/dq/actors/compute/dq_compute_actor.h>
-#include <ydb/core/formats/arrow/simple_builder/filler.h>
-#include <ydb/core/formats/arrow/simple_builder/array.h>
-#include <ydb/core/formats/arrow/simple_builder/batch.h>
 #include <ydb/core/formats/arrow/ssa_runtime_version.h>
+
+#include "helpers/get_value.h"
+#include "helpers/query_executor.h"
+#include "helpers/local.h"
+#include "helpers/writer.h"
+#include "helpers/aggregation.h"
+
+#include <ydb/core/base/tablet_pipecache.h>
 #include <ydb/core/kqp/executer_actor/kqp_executer.h>
-#include <ydb/core/tx/columnshard/data_sharing/destination/events/control.h>
-#include <ydb/core/tx/columnshard/hooks/abstract/abstract.h>
+#include <ydb/core/kqp/common/simple/kqp_event_ids.h>
+#include <ydb/core/kqp/ut/common/columnshard.h>
+#include <ydb/core/testlib/common_helper.h>
 #include <ydb/core/tx/columnshard/hooks/testing/controller.h>
-#include <ydb/core/tx/datashard/datashard.h>
 #include <ydb/core/tx/datashard/datashard_ut_common_kqp.h>
-#include <ydb/core/tx/datashard/ut_common/datashard_ut_common.h>
-#include <ydb/core/grpc_services/local_rpc/local_rpc.h>
-#include <ydb/core/grpc_services/base/base.h>
-#include <ydb/core/tx/tx_proxy/proxy.h>
-#include <ydb/core/tx/schemeshard/schemeshard.h>
-#include <ydb/core/testlib/tablet_helpers.h>
-#include <ydb/core/testlib/test_client.h>
-#include <ydb/core/testlib/cs_helper.h>
-#include <ydb/core/wrappers/fake_storage.h>
-#include <util/system/sanitizers.h>
 
-#include <fmt/format.h>
-#include <contrib/libs/apache/arrow/cpp/src/arrow/type_traits.h>
+#include <ydb/library/yql/dq/actors/dq_events_ids.h>
+#include <ydb/library/yql/dq/actors/compute/dq_compute_actor.h>
 
-namespace NKikimr {
-namespace NKqp {
+#include <library/cpp/testing/unittest/registar.h>
 
-using namespace NKikimr::NDataShard::NKqpHelpers;
+namespace NKikimr::NKqp {
 using namespace NSchemeShard;
 using namespace NActors;
 using namespace NYdb;
 using namespace NYdb::NTable;
 using namespace NYdb::NScheme;
 
-using TEvBulkUpsertRequest = NGRpcService::TGrpcRequestOperationCall<Ydb::Table::BulkUpsertRequest,
-    Ydb::Table::BulkUpsertResponse>;
-
 Y_UNIT_TEST_SUITE(KqpOlap) {
-    void PrintValue(IOutputStream& out, const NYdb::TValue& v) {
-        NYdb::TValueParser value(v);
-
-        while (value.GetKind() == NYdb::TTypeParser::ETypeKind::Optional) {
-            if (value.IsNull()) {
-                out << "<NULL>";
-                return;
-            } else {
-                value.OpenOptional();
-            }
-        }
-
-        if (value.IsNull()) {
-            out << "<NULL>";
-            return;
-        }
-
-        switch (value.GetPrimitiveType()) {
-            case NYdb::EPrimitiveType::Uint32:
-            {
-                out << value.GetUint32();
-                break;
-            }
-            case NYdb::EPrimitiveType::Uint64:
-            {
-                out << value.GetUint64();
-                break;
-            }
-            case NYdb::EPrimitiveType::Int64:
-            {
-                out << value.GetInt64();
-                break;
-            }
-            case NYdb::EPrimitiveType::Utf8:
-            {
-                out << value.GetUtf8();
-                break;
-            }
-            case NYdb::EPrimitiveType::Timestamp:
-            {
-                out << value.GetTimestamp();
-                break;
-            }
-            case NYdb::EPrimitiveType::Bool:
-            {
-                out << value.GetBool();
-                break;
-            }
-            default:
-            {
-                UNIT_ASSERT_C(false, "PrintValue not iplemented for this type");
-            }
-        }
-    }
-
-    void PrintRow(IOutputStream& out, const THashMap<TString, NYdb::TValue>& fields) {
-        for (const auto& f : fields) {
-            out << f.first << ": ";
-            PrintValue(out, f.second);
-            out << " ";
-        }
-    }
-
-    void PrintRows(IOutputStream& out, const TVector<THashMap<TString, NYdb::TValue>>& rows) {
-        for (const auto& r : rows) {
-            PrintRow(out, r);
-            out << "\n";
-        }
-    }
-
-    TVector<THashMap<TString, NYdb::TValue>> CollectRows(NYdb::NTable::TScanQueryPartIterator& it, NJson::TJsonValue* statInfo = nullptr, NJson::TJsonValue* diagnostics = nullptr) {
-        TVector<THashMap<TString, NYdb::TValue>> rows;
-        if (statInfo) {
-            *statInfo = NJson::JSON_NULL;
-        }
-        if (diagnostics) {
-            *diagnostics = NJson::JSON_NULL;
-        }
-        for (;;) {
-            auto streamPart = it.ReadNext().GetValueSync();
-            if (!streamPart.IsSuccess()) {
-                UNIT_ASSERT_C(streamPart.EOS(), streamPart.GetIssues().ToString());
-                break;
-            }
-
-            UNIT_ASSERT_C(streamPart.HasResultSet() || streamPart.HasQueryStats(),
-                "Unexpected empty scan query response.");
-
-            if (streamPart.HasQueryStats()) {
-                auto plan = streamPart.GetQueryStats().GetPlan();
-                if (plan && statInfo) {
-                    UNIT_ASSERT(NJson::ReadJsonFastTree(*plan, statInfo));
-                }
-            }
-
-            if (streamPart.HasDiagnostics()) {
-                TString diagnosticsString = streamPart.GetDiagnostics();
-                if (!diagnosticsString.empty() && diagnostics) {
-                    UNIT_ASSERT(NJson::ReadJsonFastTree(diagnosticsString, diagnostics));
-                }
-            }
-
-            if (streamPart.HasResultSet()) {
-                auto resultSet = streamPart.ExtractResultSet();
-                NYdb::TResultSetParser rsParser(resultSet);
-                while (rsParser.TryNextRow()) {
-                    THashMap<TString, NYdb::TValue> row;
-                    for (size_t ci = 0; ci < resultSet.ColumnsCount(); ++ci) {
-                        row.emplace(resultSet.GetColumnsMeta()[ci].Name, rsParser.GetValue(ci));
-                    }
-                    rows.emplace_back(std::move(row));
-                }
-            }
-        }
-        return rows;
-    }
-
-    TVector<THashMap<TString, NYdb::TValue>> ExecuteScanQuery(NYdb::NTable::TTableClient& tableClient, const TString& query, const bool verbose = true) {
-        if (verbose) {
-            Cerr << "====================================\n"
-                << "QUERY:\n" << query
-                << "\n\nRESULT:\n";
-        }
-
-        TStreamExecScanQuerySettings scanSettings;
-        auto it = tableClient.StreamExecuteScanQuery(query, scanSettings).GetValueSync();
-        auto rows = CollectRows(it);
-        if (verbose) {
-            PrintRows(Cerr, rows);
-            Cerr << "\n";
-        }
-
-        return rows;
-    }
-
-    ui64 GetUint32(const NYdb::TValue& v) {
-        NYdb::TValueParser value(v);
-        if (value.GetKind() == NYdb::TTypeParser::ETypeKind::Optional) {
-            return *value.GetOptionalUint32();
-        } else {
-            return value.GetUint32();
-        }
-    }
-
-    ui64 GetUint64(const NYdb::TValue& v) {
-        NYdb::TValueParser value(v);
-        if (value.GetKind() == NYdb::TTypeParser::ETypeKind::Optional) {
-            return *value.GetOptionalUint64();
-        } else {
-            return value.GetUint64();
-        }
-    }
-
-    TString GetUtf8(const NYdb::TValue& v) {
-        NYdb::TValueParser value(v);
-        if (value.GetKind() == NYdb::TTypeParser::ETypeKind::Optional) {
-            return *value.GetOptionalUtf8();
-        } else {
-            return value.GetUtf8();
-        }
-    }
-
-    TInstant GetTimestamp(const NYdb::TValue& v) {
-        NYdb::TValueParser value(v);
-        if (value.GetKind() == NYdb::TTypeParser::ETypeKind::Optional) {
-            return *value.GetOptionalTimestamp();
-        } else {
-            return value.GetTimestamp();
-        }
-    }
-
-    class TTypedLocalHelper: public Tests::NCS::THelper {
-    private:
-        using TBase = Tests::NCS::THelper;
-        const TString TypeName;
-        TKikimrRunner& KikimrRunner;
-        const TString TablePath;
-        const TString TableName;
-        const TString StoreName;
-    protected:
-        virtual TString GetTestTableSchema() const override {
-            TString result;
-            if (TypeName) {
-                result = R"(Columns { Name: "field" Type: ")" + TypeName + "\"}";
-            }
-            result += R"(
-                Columns { Name: "pk_int" Type: "Int64" NotNull: true }
-                Columns { Name: "ts" Type: "Timestamp" }
-                KeyColumnNames: "pk_int"
-                Engine: COLUMN_ENGINE_REPLACING_TIMESERIES
-            )";
-            return result;
-        }
-        virtual std::vector<TString> GetShardingColumns() const override {
-            return { "pk_int" };
-        }
-    public:
-        TTypedLocalHelper(const TString& typeName, TKikimrRunner& kikimrRunner, const TString& tableName = "olapTable", const TString& storeName = "olapStore")
-            : TBase(kikimrRunner.GetTestServer())
-            , TypeName(typeName)
-            , KikimrRunner(kikimrRunner)
-            , TablePath("/Root/" + storeName + "/" + tableName)
-            , TableName(tableName)
-            , StoreName(storeName)
-        {
-            SetShardingMethod("HASH_FUNCTION_CONSISTENCY_64");
-        }
-
-        void ExecuteSchemeQuery(const TString& alterQuery, const EStatus expectedStatus = EStatus::SUCCESS) const {
-            auto session = KikimrRunner.GetTableClient().CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), expectedStatus, alterResult.GetIssues().ToString());
-        }
-
-        TString GetQueryResult(const TString& request) const {
-                auto db = KikimrRunner.GetQueryClient();
-                auto result = db.ExecuteQuery(request, NYdb::NQuery::TTxControl::BeginTx().CommitTx()).ExtractValueSync();
-                UNIT_ASSERT_VALUES_EQUAL_C(result.GetStatus(), EStatus::SUCCESS, result.GetIssues().ToString());
-                const TString output = FormatResultSetYson(result.GetResultSet(0));
-                Cout << output << Endl;
-                return output;
-        }
-
-        void PrintCount() {
-            const TString selectQuery = "SELECT COUNT(*), MAX(pk_int), MIN(pk_int) FROM `" + TablePath + "`";
-
-            auto tableClient = KikimrRunner.GetTableClient();
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-            for (auto&& r : rows) {
-                for (auto&& c : r) {
-                    Cerr << c.first << ":" << Endl << c.second.GetProto().DebugString() << Endl;
-                }
-            }
-        }
-
-        class TDistribution {
-        private:
-            YDB_READONLY(ui32, Count, 0);
-            YDB_READONLY(ui32, MinCount, 0);
-            YDB_READONLY(ui32, MaxCount, 0);
-            YDB_READONLY(ui32, GroupsCount, 0);
-        public:
-            TDistribution(const ui32 count, const ui32 minCount, const ui32 maxCount, const ui32 groupsCount)
-                : Count(count)
-                , MinCount(minCount)
-                , MaxCount(maxCount)
-                , GroupsCount(groupsCount)
-            {
-
-            }
-
-            TString DebugString() const {
-                return TStringBuilder()
-                    << "count=" << Count << ";"
-                    << "min_count=" << MinCount << ";"
-                    << "max_count=" << MaxCount << ";"
-                    << "groups_count=" << GroupsCount << ";";
-            }
-        };
-
-        TDistribution GetDistribution(const bool verbose = false) {
-            const TString selectQuery = "PRAGMA Kikimr.OptUseFinalizeByKey='true';SELECT COUNT(*) as c, field FROM `" + TablePath + "` GROUP BY field ORDER BY field";
-
-            auto tableClient = KikimrRunner.GetTableClient();
-            auto rows = ExecuteScanQuery(tableClient, selectQuery, verbose);
-            ui32 count = 0;
-            std::optional<ui32> minCount;
-            std::optional<ui32> maxCount;
-            std::set<TString> groups;
-            for (auto&& r : rows) {
-                for (auto&& c : r) {
-                    if (c.first == "c") {
-                        const ui64 v = GetUint64(c.second);
-                        count += v;
-                        if (!minCount || *minCount > v) {
-                            minCount = v;
-                        }
-                        if (!maxCount || *maxCount < v) {
-                            maxCount = v;
-                        }
-                    } else if (c.first == "field") {
-                        Y_ABORT_UNLESS(groups.emplace(c.second.GetProto().DebugString()).second);
-                    }
-                    if (verbose) {
-                        Cerr << c.first << ":" << Endl << c.second.GetProto().DebugString() << Endl;
-                    }
-                }
-            }
-            Y_ABORT_UNLESS(maxCount);
-            Y_ABORT_UNLESS(minCount);
-            return TDistribution(count, *minCount, *maxCount, groups.size());
-        }
-
-        void GetVolumes(ui64& rawBytes, ui64& bytes, const bool verbose = false, const std::vector<TString> columnNames = {}) {
-            TString selectQuery = "SELECT * FROM `" + TablePath + "/.sys/primary_index_stats` WHERE Activity = true";
-            if (columnNames.size()) {
-                selectQuery += " AND EntityName IN ('" + JoinSeq("','", columnNames) + "')";
-            }
-
-            auto tableClient = KikimrRunner.GetTableClient();
-
-            std::optional<ui64> rawBytesPred;
-            std::optional<ui64> bytesPred;
-            while (true) {
-                auto rows = ExecuteScanQuery(tableClient, selectQuery);
-                rawBytes = 0;
-                bytes = 0;
-                for (auto&& r : rows) {
-                    if (verbose) {
-                        Cerr << "-------" << Endl;
-                    }
-                    for (auto&& c : r) {
-                        if (c.first == "RawBytes") {
-                            rawBytes += GetUint64(c.second);
-                        }
-                        if (c.first == "BlobRangeSize") {
-                            bytes += GetUint64(c.second);
-                        }
-                        if (verbose) {
-                            Cerr << c.first << ":" << Endl << c.second.GetProto().DebugString() << Endl;
-                        }
-                    }
-                }
-                if (rawBytesPred && *rawBytesPred == rawBytes && bytesPred && *bytesPred == bytes) {
-                    break;
-                } else {
-                    rawBytesPred = rawBytes;
-                    bytesPred = bytes;
-                    Cerr << "Wait changes: " << bytes << "/" << rawBytes << Endl;
-                    Sleep(TDuration::Seconds(5));
-                }
-            }
-            Cerr << bytes << "/" << rawBytes << Endl;
-        }
-
-        void GetStats(std::vector<NKikimrColumnShardStatisticsProto::TPortionStorage>& stats, const bool verbose = false) {
-            TString selectQuery = "SELECT * FROM `" + TablePath + "/.sys/primary_index_portion_stats` WHERE Activity = true";
-            auto tableClient = KikimrRunner.GetTableClient();
-            auto rows = ExecuteScanQuery(tableClient, selectQuery, verbose);
-            for (auto&& r : rows) {
-                for (auto&& c : r) {
-                    if (c.first == "Stats") {
-                        NKikimrColumnShardStatisticsProto::TPortionStorage store;
-                        AFL_VERIFY(google::protobuf::TextFormat::ParseFromString(GetUtf8(c.second), &store));
-                        stats.emplace_back(store);
-                    }
-                }
-            }
-
-        }
-
-        void GetCount(ui64& count) {
-            const TString selectQuery = "SELECT COUNT(*) as a FROM `" + TablePath + "`";
-
-            auto tableClient = KikimrRunner.GetTableClient();
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-            for (auto&& r : rows) {
-                for (auto&& c : r) {
-                    if (c.first == "a") {
-                        count = GetUint64(c.second);
-                    }
-                }
-            }
-        }
-
-        template <class TFiller>
-        void FillTable(const TFiller& fillPolicy, const ui32 pkKff = 0, const ui32 numRows = 800000) const {
-            std::vector<NArrow::NConstruction::IArrayBuilder::TPtr> builders;
-            builders.emplace_back(NArrow::NConstruction::TSimpleArrayConstructor<NArrow::NConstruction::TIntSeqFiller<arrow::Int64Type>>::BuildNotNullable("pk_int", numRows * pkKff));
-            builders.emplace_back(std::make_shared<NArrow::NConstruction::TSimpleArrayConstructor<TFiller>>("field", fillPolicy));
-            NArrow::NConstruction::TRecordBatchConstructor batchBuilder(builders);
-            std::shared_ptr<arrow::RecordBatch> batch = batchBuilder.BuildBatch(numRows);
-            TBase::SendDataViaActorSystem(TablePath, batch);
-        }
-
-        void FillPKOnly(const double pkKff = 0, const ui32 numRows = 800000) const {
-            std::vector<NArrow::NConstruction::IArrayBuilder::TPtr> builders;
-            builders.emplace_back(NArrow::NConstruction::TSimpleArrayConstructor<NArrow::NConstruction::TIntSeqFiller<arrow::Int64Type>>::BuildNotNullable("pk_int", numRows * pkKff));
-            NArrow::NConstruction::TRecordBatchConstructor batchBuilder(builders);
-            std::shared_ptr<arrow::RecordBatch> batch = batchBuilder.BuildBatch(numRows);
-            TBase::SendDataViaActorSystem(TablePath, batch);
-        }
-
-        void CreateTestOlapTable(ui32 storeShardsCount = 4, ui32 tableShardsCount = 3) {
-            CreateOlapTableWithStore(TableName, StoreName, storeShardsCount, tableShardsCount);
-        }
-    };
-
-    class TLocalHelper: public Tests::NCS::THelper {
-    private:
-        using TBase = Tests::NCS::THelper;
-    public:
-        TLocalHelper& SetShardingMethod(const TString& value) {
-            TBase::SetShardingMethod(value);
-            return *this;
-        }
-
-        void CreateTestOlapTable(TString tableName = "olapTable", TString storeName = "olapStore",
-            ui32 storeShardsCount = 4, ui32 tableShardsCount = 3) {
-            CreateOlapTableWithStore(tableName, storeName, storeShardsCount, tableShardsCount);
-        }
-        using TBase::TBase;
-
-        TLocalHelper(TKikimrRunner& runner)
-            : TBase(runner.GetTestServer()) {
-
-        }
-    };
 
     class TExtLocalHelper: public TLocalHelper {
     private:
@@ -510,86 +82,6 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
 
         }
     };
-
-    class TClickHelper : public Tests::NCS::TCickBenchHelper {
-    private:
-        using TBase = Tests::NCS::TCickBenchHelper;
-    public:
-        using TBase::TBase;
-
-        TClickHelper(TKikimrRunner& runner)
-            : TBase(runner.GetTestServer())
-        {}
-
-        void CreateClickBenchTable(TString tableName = "benchTable", ui32 shardsCount = 4) {
-            TActorId sender = Server.GetRuntime()->AllocateEdgeActor();
-
-            TBase::CreateTestOlapTable(sender, "", Sprintf(R"(
-                Name: "%s"
-                ColumnShardCount: %d
-                Schema {
-                    %s
-                }
-                Sharding {
-                    HashSharding {
-                        Function: HASH_FUNCTION_CONSISTENCY_64
-                        Columns: "EventTime"
-                    }
-                })", tableName.c_str(), shardsCount, PROTO_SCHEMA));
-        }
-    };
-
-    class TTableWithNullsHelper : public Tests::NCS::TTableWithNullsHelper {
-    private:
-        using TBase = Tests::NCS::TTableWithNullsHelper;
-    public:
-        using TBase::TBase;
-
-        TTableWithNullsHelper(TKikimrRunner& runner)
-            : TBase(runner.GetTestServer())
-        {}
-
-        void CreateTableWithNulls(TString tableName = "tableWithNulls", ui32 shardsCount = 4) {
-            TActorId sender = Server.GetRuntime()->AllocateEdgeActor();
-
-            TBase::CreateTestOlapTable(sender, "", Sprintf(R"(
-                Name: "%s"
-                ColumnShardCount: %d
-                Schema {
-                    %s
-                }
-                Sharding {
-                    HashSharding {
-                        Function: HASH_FUNCTION_CONSISTENCY_64
-                        Columns: "id"
-                    }
-                })", tableName.c_str(), shardsCount, PROTO_SCHEMA));
-        }
-    };
-
-    void WriteTestData(TKikimrRunner& kikimr, TString testTable, ui64 pathIdBegin, ui64 tsBegin, size_t rowCount, bool withSomeNulls = false) {
-        UNIT_ASSERT(testTable != "/Root/benchTable"); // TODO: check schema instead
-        TLocalHelper lHelper(kikimr);
-        if (withSomeNulls) {
-            lHelper.WithSomeNulls();
-        }
-        auto batch = lHelper.TestArrowBatch(pathIdBegin, tsBegin, rowCount);
-        lHelper.SendDataViaActorSystem(testTable, batch);
-    }
-
-    void WriteTestDataForClickBench(TKikimrRunner& kikimr, TString testTable, ui64 pathIdBegin, ui64 tsBegin, size_t rowCount) {
-        UNIT_ASSERT(testTable == "/Root/benchTable"); // TODO: check schema instead
-        TClickHelper lHelper(kikimr.GetTestServer());
-        auto batch = lHelper.TestArrowBatch(pathIdBegin, tsBegin, rowCount);
-        lHelper.SendDataViaActorSystem(testTable, batch);
-    }
-
-    void WriteTestDataForTableWithNulls(TKikimrRunner& kikimr, TString testTable) {
-        UNIT_ASSERT(testTable == "/Root/tableWithNulls"); // TODO: check schema instead
-        TTableWithNullsHelper lHelper(kikimr.GetTestServer());
-        auto batch = lHelper.TestArrowBatch();
-        lHelper.SendDataViaActorSystem(testTable, batch);
-    }
 
     void CreateTableOfAllTypes(TKikimrRunner& kikimr) {
         auto& legacyClient = kikimr.GetTestClient();
@@ -724,59 +216,6 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
             },
 #endif
         };
-    }
-
-    template <typename TClient>
-    auto StreamExplainQuery(const TString& query, TClient& client) {
-        if constexpr (std::is_same_v<NYdb::NTable::TTableClient, TClient>) {
-            TStreamExecScanQuerySettings scanSettings;
-            scanSettings.Explain(true);
-            return client.StreamExecuteScanQuery(query, scanSettings).GetValueSync();
-        } else {
-            NYdb::NQuery::TExecuteQuerySettings scanSettings;
-            scanSettings.ExecMode(NYdb::NQuery::EExecMode::Explain);
-            return client.StreamExecuteQuery(query, NYdb::NQuery::TTxControl::BeginTx().CommitTx(), scanSettings).GetValueSync();
-        }
-    }
-
-    template <typename TClient>
-    void CheckPlanForAggregatePushdown(
-        const TString& query,
-        TClient& client,
-        const std::vector<std::string>& expectedPlanNodes,
-        const std::string& readNodeType)
-    {
-        auto res = StreamExplainQuery(query, client);
-        UNIT_ASSERT_C(res.IsSuccess(), res.GetIssues().ToString());
-
-        auto planRes = CollectStreamResult(res);
-        auto ast = planRes.QueryStats->Getquery_ast();
-        Cerr << "JSON Plan:" << Endl;
-        Cerr << planRes.PlanJson.GetOrElse("NO_PLAN") << Endl;
-        Cerr << "AST:" << Endl;
-        Cerr << ast << Endl;
-        for (auto planNode : expectedPlanNodes) {
-            UNIT_ASSERT_C(ast.find(planNode) != std::string::npos,
-                TStringBuilder() << planNode << " was not found. Query: " << query);
-        }
-        UNIT_ASSERT_C(ast.find("SqueezeToDict") == std::string::npos, TStringBuilder() << "SqueezeToDict denied for aggregation requests. Query: " << query);
-
-        if (!readNodeType.empty()) {
-            NJson::TJsonValue planJson;
-            NJson::ReadJsonTree(*planRes.PlanJson, &planJson, true);
-            auto readNode = FindPlanNodeByKv(planJson, "Node Type", readNodeType.c_str());
-            UNIT_ASSERT(readNode.IsDefined());
-
-            auto& operators = readNode.GetMapSafe().at("Operators").GetArraySafe();
-            for (auto& op : operators) {
-                if (op.GetMapSafe().at("Name") == "TableFullScan") {
-                    auto ssaProgram = op.GetMapSafe().at("SsaProgram");
-                    UNIT_ASSERT(ssaProgram.IsDefined());
-                    UNIT_ASSERT(FindPlanNodes(ssaProgram, "Projection").size());
-                    break;
-                }
-            }
-        }
     }
 
     Y_UNIT_TEST(SimpleQueryOlap) {
@@ -1205,423 +644,6 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
 
         UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
         CompareYson(StreamResultToYson(it), "[]");
-    }
-
-    Y_UNIT_TEST(Aggregation) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-
-        TLocalHelper(kikimr).CreateTestOlapTable();
-
-        auto tableClient = kikimr.GetTableClient();
-
-        {
-            auto it = tableClient.StreamExecuteScanQuery(R"(
-                --!syntax_v1
-
-                SELECT
-                    COUNT(*)
-                FROM `/Root/olapStore/olapTable`
-            )").GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cout << result << Endl;
-            CompareYson(result, R"([[0u;]])");
-        }
-
-        {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 10000, 3000000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 11000, 3001000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 12000, 3002000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 13000, 3003000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 14000, 3004000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 20000, 2000000, 7000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 30000, 1000000, 11000);
-        }
-
-        {
-            auto it = tableClient.StreamExecuteScanQuery(R"(
-                --!syntax_v1
-
-                SELECT
-                    COUNT(*), MAX(`resource_id`), MAX(`timestamp`), MIN(LENGTH(`message`))
-                FROM `/Root/olapStore/olapTable`
-            )").GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cout << result << Endl;
-            CompareYson(result, R"([[23000u;["40999"];[3004999u];[1036u]]])");
-        }
-
-        {
-            auto it = tableClient.StreamExecuteScanQuery(R"(
-                --!syntax_v1
-
-                SELECT
-                    COUNT(*)
-                FROM `/Root/olapStore/olapTable`
-            )").GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cout << result << Endl;
-            CompareYson(result, R"([[23000u;]])");
-        }
-    }
-
-    Y_UNIT_TEST(IndexesActualization) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        csController->SetPeriodicWakeupActivationPeriod(TDuration::Seconds(1));
-        csController->SetLagForCompactionBeforeTierings(TDuration::Seconds(1));
-
-        TLocalHelper(kikimr).CreateTestOlapTable();
-        auto tableClient = kikimr.GetTableClient();
-
-        Tests::NCommon::TLoggerInit(kikimr).SetComponents({NKikimrServices::TX_COLUMNSHARD}, "CS").SetPriority(NActors::NLog::PRI_DEBUG).Initialize();
-
-        std::vector<TString> uids;
-        std::vector<TString> resourceIds;
-        std::vector<ui32> levels;
-
-        {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1000000, 300000000, 10000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1100000, 300100000, 10000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1200000, 300200000, 10000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1300000, 300300000, 10000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1400000, 300400000, 10000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 2000000, 200000000, 70000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 3000000, 100000000, 110000);
-
-            const auto filler = [&](const ui32 startRes, const ui32 startUid, const ui32 count) {
-                for (ui32 i = 0; i < count; ++i) {
-                    uids.emplace_back("uid_" + ::ToString(startUid + i));
-                    resourceIds.emplace_back(::ToString(startRes + i));
-                    levels.emplace_back(i % 5);
-                }
-            };
-
-            filler(1000000, 300000000, 10000);
-            filler(1100000, 300100000, 10000);
-            filler(1200000, 300200000, 10000);
-            filler(1300000, 300300000, 10000);
-            filler(1400000, 300400000, 10000);
-            filler(2000000, 200000000, 70000);
-            filler(3000000, 100000000, 110000);
-
-        }
-
-        {
-            auto alterQuery = TStringBuilder() <<
-                R"(ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_INDEX, NAME=index_uid, TYPE=BLOOM_FILTER, 
-                    FEATURES=`{"column_names" : ["uid"], "false_positive_probability" : 0.05}`);
-                )";
-            auto session = tableClient.CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-        }
-        {
-            auto alterQuery = TStringBuilder() <<
-                R"(ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_INDEX, NAME=index_resource_id, TYPE=BLOOM_FILTER, 
-                    FEATURES=`{"column_names" : ["resource_id", "level"], "false_positive_probability" : 0.05}`);
-                )";
-            auto session = tableClient.CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-        }
-
-        {
-            auto alterQuery = TStringBuilder() <<
-                "ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_OPTIONS, SCHEME_NEED_ACTUALIZATION=`true`);";
-            auto session = tableClient.CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-        }
-        csController->WaitActualization(TDuration::Seconds(10));
-        {
-            auto it = tableClient.StreamExecuteScanQuery(R"(
-                --!syntax_v1
-
-                SELECT
-                    COUNT(*)
-                FROM `/Root/olapStore/olapTable`
-                WHERE ((resource_id = '2' AND level = 222222) OR (resource_id = '1' AND level = 111111) OR (resource_id LIKE '%11dd%')) AND uid = '222'
-            )").GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cerr << result << Endl;
-            Cerr << csController->GetIndexesSkippingOnSelect().Val() << " / " << csController->GetIndexesApprovedOnSelect().Val() << Endl;
-            CompareYson(result, R"([[0u;]])");
-            AFL_VERIFY(csController->GetIndexesSkippedNoData().Val() == 0);
-            AFL_VERIFY(csController->GetIndexesApprovedOnSelect().Val() < csController->GetIndexesSkippingOnSelect().Val() * 0.4)
-                ("approve", csController->GetIndexesApprovedOnSelect().Val())("skip", csController->GetIndexesSkippingOnSelect().Val());
-        }
-    }
-
-    Y_UNIT_TEST(IndexesActualizationRebuildScheme) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        csController->SetPeriodicWakeupActivationPeriod(TDuration::Seconds(1));
-
-        TLocalHelper(kikimr).CreateTestOlapTable();
-        auto tableClient = kikimr.GetTableClient();
-
-        std::vector<TString> uids;
-        std::vector<TString> resourceIds;
-        std::vector<ui32> levels;
-
-        {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1000000, 300000000, 10000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1100000, 300100000, 10000);
-
-            const auto filler = [&](const ui32 startRes, const ui32 startUid, const ui32 count) {
-                for (ui32 i = 0; i < count; ++i) {
-                    uids.emplace_back("uid_" + ::ToString(startUid + i));
-                    resourceIds.emplace_back(::ToString(startRes + i));
-                    levels.emplace_back(i % 5);
-                }
-            };
-
-            filler(1000000, 300000000, 10000);
-            filler(1100000, 300100000, 10000);
-
-        }
-
-        for (ui32 i = 0; i < 10; ++i) {
-            auto alterQuery = TStringBuilder() <<
-                "ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_OPTIONS, SCHEME_NEED_ACTUALIZATION=`true`);";
-            auto session = tableClient.CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-        }
-        const ui64 startCount = csController->GetActualizationRefreshSchemeCount().Val();
-        AFL_VERIFY(startCount == 30);
-
-        for (auto&& i : csController->GetShardActualIds()) {
-            kikimr.GetTestServer().GetRuntime()->Send(MakePipePeNodeCacheID(false), NActors::TActorId(), new TEvPipeCache::TEvForward(
-                new TEvents::TEvPoisonPill(), i, false));
-        }
-
-        {
-            auto it = tableClient.StreamExecuteScanQuery(R"(
-                --!syntax_v1
-                SELECT
-                    COUNT(*)
-                FROM `/Root/olapStore/olapTable`
-            )").GetValueSync();
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cout << result << Endl;
-            CompareYson(result, R"([[20000u;]])");
-        }
-
-        AFL_VERIFY(startCount + 3 /*tables count*/ * 2 /*normalizers + main_load*/ == 
-            (ui64)csController->GetActualizationRefreshSchemeCount().Val())("start", startCount)("count", csController->GetActualizationRefreshSchemeCount().Val());
-    }
-
-    Y_UNIT_TEST(Indexes) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-
-        TLocalHelper(kikimr).CreateTestOlapTable();
-        auto tableClient = kikimr.GetTableClient();
-
-//        Tests::NCommon::TLoggerInit(kikimr).Initialize();
-
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-
-        {
-            auto alterQuery = TStringBuilder() <<
-                R"(ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_INDEX, NAME=index_uid, TYPE=BLOOM_FILTER, 
-                    FEATURES=`{"column_names" : ["uid"], "false_positive_probability" : 0.05}`);
-                )";
-            auto session = tableClient.CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-        }
-        {
-            auto alterQuery = TStringBuilder() <<
-                R"(ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_INDEX, NAME=index_resource_id, TYPE=BLOOM_FILTER, 
-                    FEATURES=`{"column_names" : ["resource_id", "level"], "false_positive_probability" : 0.05}`);
-                )";
-            auto session = tableClient.CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-        }
-
-        std::vector<TString> uids;
-        std::vector<TString> resourceIds;
-        std::vector<ui32> levels;
-
-        {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1000000, 300000000, 10000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1100000, 300100000, 10000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1200000, 300200000, 10000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1300000, 300300000, 10000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 1400000, 300400000, 10000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 2000000, 200000000, 70000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 3000000, 100000000, 110000);
-
-            const auto filler = [&](const ui32 startRes, const ui32 startUid, const ui32 count) {
-                for (ui32 i = 0; i < count; ++i) {
-                    uids.emplace_back("uid_" + ::ToString(startUid + i));
-                    resourceIds.emplace_back(::ToString(startRes + i));
-                    levels.emplace_back(i % 5);
-                }
-            };
-
-            filler(1000000, 300000000, 10000);
-            filler(1100000, 300100000, 10000);
-            filler(1200000, 300200000, 10000);
-            filler(1300000, 300300000, 10000);
-            filler(1400000, 300400000, 10000);
-            filler(2000000, 200000000, 70000);
-            filler(3000000, 100000000, 110000);
-            
-        }
-
-        {
-            auto it = tableClient.StreamExecuteScanQuery(R"(
-                --!syntax_v1
-
-                SELECT
-                    COUNT(*)
-                FROM `/Root/olapStore/olapTable`
-            )").GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cout << result << Endl;
-            CompareYson(result, R"([[230000u;]])");
-        }
-
-        AFL_VERIFY(csController->GetIndexesSkippingOnSelect().Val() == 0);
-        AFL_VERIFY(csController->GetIndexesApprovedOnSelect().Val() == 0);
-        TInstant start = Now();
-        ui32 compactionsStart = csController->GetCompactions().Val();
-        while (Now() - start < TDuration::Seconds(10)) {
-            if (compactionsStart != csController->GetCompactions().Val()) {
-                compactionsStart = csController->GetCompactions().Val();
-                start = Now();
-            }
-            Cerr << "WAIT_COMPACTION: " << csController->GetCompactions().Val() << Endl;
-            Sleep(TDuration::Seconds(1));
-        }
-
-        {
-            auto it = tableClient.StreamExecuteScanQuery(R"(
-                --!syntax_v1
-
-                SELECT
-                    COUNT(*)
-                FROM `/Root/olapStore/olapTable`
-                WHERE ((resource_id = '2' AND level = 222222) OR (resource_id = '1' AND level = 111111) OR (resource_id LIKE '%11dd%')) AND uid = '222'
-            )").GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cout << result << Endl;
-            Cout << csController->GetIndexesSkippingOnSelect().Val() << " / " << csController->GetIndexesApprovedOnSelect().Val() << Endl;
-            CompareYson(result, R"([[0u;]])");
-            AFL_VERIFY(csController->GetIndexesSkippedNoData().Val() == 0);
-            AFL_VERIFY(csController->GetIndexesApprovedOnSelect().Val() < csController->GetIndexesSkippingOnSelect().Val() * 0.3);
-        }
-        ui32 requestsCount = 100;
-        for (ui32 i = 0; i < requestsCount; ++i) {
-            const ui32 idx = RandomNumber<ui32>(uids.size());
-            const auto query = [](const TString& res, const TString& uid, const ui32 level) {
-                TStringBuilder sb;
-                sb << "SELECT" << Endl;
-                sb << "COUNT(*)" << Endl;
-                sb << "FROM `/Root/olapStore/olapTable`" << Endl;
-                sb << "WHERE(" << Endl;
-                sb << "resource_id = '" << res << "' AND" << Endl;
-                sb << "uid= '" << uid << "' AND" << Endl;
-                sb << "level= " << level << Endl;
-                sb << ")";
-                return sb;
-            };
-            auto it = tableClient.StreamExecuteScanQuery(query(resourceIds[idx], uids[idx], levels[idx])).GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cout << csController->GetIndexesSkippingOnSelect().Val() << " / " << csController->GetIndexesApprovedOnSelect().Val() << " / " << csController->GetIndexesSkippedNoData().Val() << Endl;
-            CompareYson(result, R"([[1u;]])");
-        }
-
-        AFL_VERIFY(csController->GetIndexesApprovedOnSelect().Val() < 0.20 * csController->GetIndexesSkippingOnSelect().Val());
-
-    }
-
-    Y_UNIT_TEST(IndexesModificationError) {
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-
-        TLocalHelper(kikimr).CreateTestOlapTable();
-        auto tableClient = kikimr.GetTableClient();
-
-        //        Tests::NCommon::TLoggerInit(kikimr).Initialize();
-
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-
-        {
-            auto alterQuery = TStringBuilder() <<
-                R"(ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_INDEX, NAME=index_uid, TYPE=BLOOM_FILTER, 
-                    FEATURES=`{"column_names" : ["uid"], "false_positive_probability" : 0.05}`);
-                )";
-            auto session = tableClient.CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-        }
-
-        {
-            auto alterQuery = TStringBuilder() <<
-                R"(ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_INDEX, NAME=index_uid, TYPE=BLOOM_FILTER, 
-                    FEATURES=`{"column_names" : ["uid", "resource_id"], "false_positive_probability" : 0.05}`);
-                )";
-            auto session = tableClient.CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_UNEQUAL(alterResult.GetStatus(), EStatus::SUCCESS);
-        }
-
-        {
-            auto alterQuery = TStringBuilder() <<
-                R"(ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_INDEX, NAME=index_uid, TYPE=BLOOM_FILTER, 
-                    FEATURES=`{"column_names" : ["uid"], "false_positive_probability" : 0.005}`);
-                )";
-            auto session = tableClient.CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_UNEQUAL(alterResult.GetStatus(), EStatus::SUCCESS);
-        }
-
-        {
-            auto alterQuery = TStringBuilder() <<
-                R"(ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_INDEX, NAME=index_uid, TYPE=BLOOM_FILTER, 
-                    FEATURES=`{"column_names" : ["uid"], "false_positive_probability" : 0.01}`);
-                )";
-            auto session = tableClient.CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-        }
-
-        {
-            auto alterQuery = TStringBuilder() << "ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=DROP_INDEX, NAME=index_uid);";
-            auto session = tableClient.CreateSession().GetValueSync().GetSession();
-            auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-            UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-        }
-
     }
 
     Y_UNIT_TEST(PushdownFilter) {
@@ -2301,2178 +1323,6 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
                           TStringBuilder() << "NarrowMap was removed. Query: " << query);
     }
 
-    Y_UNIT_TEST(AggregationCountPushdown) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false)
-            .SetForceColumnTablesCompositeMarks(true);
-        TKikimrRunner kikimr(settings);
-
-        TLocalHelper(kikimr).CreateTestOlapTable();
-        auto tableClient = kikimr.GetTableClient();
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-
-        {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 10000, 3000000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 11000, 3001000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 12000, 3002000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 13000, 3003000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 14000, 3004000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 20000, 2000000, 7000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 30000, 1000000, 11000);
-        }
-        while (csController->GetIndexations().Val() == 0) {
-            Cout << "Wait indexation..." << Endl;
-            Sleep(TDuration::Seconds(2));
-        }
-        AFL_VERIFY(Singleton<NWrappers::NExternalStorage::TFakeExternalStorage>()->GetSize());
-
-        {
-            TString query = R"(
-                --!syntax_v1
-                SELECT
-                    COUNT(level)
-                FROM `/Root/olapStore/olapTable`
-            )";
-            auto opStartTime = Now();
-            auto it = tableClient.StreamExecuteScanQuery(query).GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cerr << "!!!\nPushdown query execution time: " << (Now() - opStartTime).MilliSeconds() << "\n!!!\n";
-            Cout << result << Endl;
-            CompareYson(result, R"([[23000u;]])");
-
-            // Check plan
-#if SSA_RUNTIME_VERSION >= 2U
-            CheckPlanForAggregatePushdown(query, tableClient, { "TKqpOlapAgg" }, "TableFullScan");
-#else
-            CheckPlanForAggregatePushdown(query, tableClient, { "CombineCore" }, "");
-#endif
-        }
-    }
-
-    Y_UNIT_TEST(AggregationCountGroupByPushdown) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false)
-            .SetForceColumnTablesCompositeMarks(true);
-        TKikimrRunner kikimr(settings);
-
-        TLocalHelper(kikimr).CreateTestOlapTable();
-        auto tableClient = kikimr.GetTableClient();
-
-        {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 10000, 3000000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 11000, 3001000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 12000, 3002000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 13000, 3003000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 14000, 3004000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 20000, 2000000, 7000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 30000, 1000000, 11000);
-        }
-
-        {
-            TString query = R"(
-                --!syntax_v1
-                PRAGMA Kikimr.OptUseFinalizeByKey;
-                SELECT
-                    level, COUNT(level)
-                FROM `/Root/olapStore/olapTable`
-                GROUP BY level
-                ORDER BY level
-            )";
-            auto it = tableClient.StreamExecuteScanQuery(query).GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cout << result << Endl;
-            CompareYson(result, R"([[[0];4600u];[[1];4600u];[[2];4600u];[[3];4600u];[[4];4600u]])");
-
-            // Check plan
-#if SSA_RUNTIME_VERSION >= 2U
-            CheckPlanForAggregatePushdown(query, tableClient, { "WideCombiner" }, "Aggregate-TableFullScan");
-//            CheckPlanForAggregatePushdown(query, tableClient, { "TKqpOlapAgg" }, "TableFullScan");
-#else
-            CheckPlanForAggregatePushdown(query, tableClient, { "CombineCore" }, "");
-#endif
-        }
-    }
-
-    Y_UNIT_TEST_TWIN(CountAllPushdown, UseLlvm) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false)
-            .SetForceColumnTablesCompositeMarks(true);
-        TKikimrRunner kikimr(settings);
-
-        TLocalHelper(kikimr).CreateTestOlapTable();
-        auto tableClient = kikimr.GetTableClient();
-
-        {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 10000, 3000000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 11000, 3001000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 12000, 3002000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 13000, 3003000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 14000, 3004000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 20000, 2000000, 7000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 30000, 1000000, 11000);
-        }
-
-        {
-            TString query = fmt::format(R"(
-                --!syntax_v1
-                PRAGMA ydb.UseLlvm = "{}";
-
-                SELECT
-                    COUNT(*)
-                FROM `/Root/olapStore/olapTable`
-            )", UseLlvm ? "true" : "false");
-            auto it = tableClient.StreamExecuteScanQuery(query).GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cout << result << Endl;
-            CompareYson(result, R"([[23000u;]])");
-
-            // Check plan
-#if SSA_RUNTIME_VERSION >= 2U
-            CheckPlanForAggregatePushdown(query, tableClient, { "TKqpOlapAgg" }, "TableFullScan");
-#else
-            CheckPlanForAggregatePushdown(query, tableClient, { "Condense" }, "");
-#endif
-        }
-    }
-
-    Y_UNIT_TEST_TWIN(CountAllPushdownBackwardCompatibility, EnableLlvm) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false)
-            .SetForceColumnTablesCompositeMarks(true);
-        TKikimrRunner kikimr(settings);
-
-        TLocalHelper(kikimr).CreateTestOlapTable();
-        auto tableClient = kikimr.GetTableClient();
-
-        {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 10000, 3000000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 11000, 3001000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 12000, 3002000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 13000, 3003000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 14000, 3004000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 20000, 2000000, 7000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 30000, 1000000, 11000);
-        }
-
-        {
-            TString query = fmt::format(R"(
-                --!syntax_v1
-                PRAGMA Kikimr.EnableLlvm = "{}";
-
-                SELECT
-                    COUNT(*)
-                FROM `/Root/olapStore/olapTable`
-            )", EnableLlvm ? "true" : "false");
-            auto it = tableClient.StreamExecuteScanQuery(query).GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cout << result << Endl;
-            CompareYson(result, R"([[23000u;]])");
-
-            // Check plan
-#if SSA_RUNTIME_VERSION >= 2U
-            CheckPlanForAggregatePushdown(query, tableClient, { "TKqpOlapAgg" }, "TableFullScan");
-#else
-            CheckPlanForAggregatePushdown(query, tableClient, { "Condense" }, "");
-#endif
-        }
-    }
-
-    Y_UNIT_TEST(CountAllNoPushdown) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false)
-            .SetForceColumnTablesCompositeMarks(true);
-        TKikimrRunner kikimr(settings);
-
-        TLocalHelper(kikimr).CreateTestOlapTable();
-        auto tableClient = kikimr.GetTableClient();
-
-        {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 10000, 3000000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 11000, 3001000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 12000, 3002000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 13000, 3003000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 14000, 3004000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 20000, 2000000, 7000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 30000, 1000000, 11000);
-        }
-
-        {
-            auto it = tableClient.StreamExecuteScanQuery(R"(
-                --!syntax_v1
-                SELECT
-                    COUNT(*)
-                FROM `/Root/olapStore/olapTable`
-            )").GetValueSync();
-
-            UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-            TString result = StreamResultToYson(it);
-            Cout << result << Endl;
-            CompareYson(result, R"([[23000u;]])");
-        }
-    }
-
-    class TExpectedLimitChecker {
-    private:
-        std::optional<ui32> ExpectedLimit;
-        std::optional<ui32> ExpectedResultCount;
-        ui32 CheckScanData = 0;
-        ui32 CheckScanTask = 0;
-    public:
-        TExpectedLimitChecker& SetExpectedLimit(const ui32 value) {
-            ExpectedLimit = value;
-            ExpectedResultCount = value;
-            return *this;
-        }
-        TExpectedLimitChecker& SetExpectedResultCount(const ui32 value) {
-            ExpectedResultCount = value;
-            return *this;
-        }
-        bool CheckExpectedLimitOnScanData(const ui32 resultCount) {
-            if (!ExpectedResultCount) {
-                return true;
-            }
-            ++CheckScanData;
-            UNIT_ASSERT_LE(resultCount, *ExpectedResultCount);
-            return true;
-        }
-        bool CheckExpectedLimitOnScanTask(const ui32 taskLimit) {
-            if (!ExpectedLimit) {
-                return true;
-            }
-            ++CheckScanTask;
-            UNIT_ASSERT_EQUAL(taskLimit, *ExpectedLimit);
-            return true;
-        }
-        bool CheckFinish() const {
-            if (!ExpectedLimit) {
-                return true;
-            }
-            return CheckScanData && CheckScanTask;
-        }
-    };
-
-    class TExpectedRecordChecker {
-    private:
-        std::optional<ui32> ExpectedColumnsCount;
-        ui32 CheckScanData = 0;
-    public:
-        TExpectedRecordChecker& SetExpectedColumnsCount(const ui32 value) {
-            ExpectedColumnsCount = value;
-            return *this;
-        }
-        bool CheckExpectedOnScanData(const ui32 columnsCount) {
-            if (!ExpectedColumnsCount) {
-                return true;
-            }
-            ++CheckScanData;
-            UNIT_ASSERT_EQUAL(columnsCount, *ExpectedColumnsCount);
-            return true;
-        }
-        bool CheckFinish() const {
-            if (!ExpectedColumnsCount) {
-                return true;
-            }
-            return CheckScanData;
-        }
-    };
-
-    class TAggregationTestCase {
-    private:
-        TString Query;
-        TString ExpectedReply;
-        std::vector<std::string> ExpectedPlanOptions;
-        bool Pushdown = true;
-        std::string ExpectedReadNodeType;
-        TExpectedLimitChecker LimitChecker;
-        TExpectedRecordChecker RecordChecker;
-        bool UseLlvm = true;
-    public:
-        void FillExpectedAggregationGroupByPlanOptions() {
-#if SSA_RUNTIME_VERSION >= 2U
-//            AddExpectedPlanOptions("TKqpOlapAgg");
-            AddExpectedPlanOptions("WideCombiner");
-#else
-            AddExpectedPlanOptions("CombineCore");
-#endif
-        }
-        TString GetFixedQuery() const {
-            TStringBuilder queryFixed;
-            queryFixed << "--!syntax_v1" << Endl;
-            if (!Pushdown) {
-                queryFixed << "PRAGMA Kikimr.OptEnableOlapPushdown = \"false\";" << Endl;
-            }
-            if (!UseLlvm) {
-                queryFixed << "PRAGMA Kikimr.UseLlvm = \"false\";" << Endl;
-            }
-            queryFixed << "PRAGMA Kikimr.OptUseFinalizeByKey;" << Endl;
-
-            queryFixed << Query << Endl;
-            Cerr << "REQUEST:\n" << queryFixed << Endl;
-            return queryFixed;
-        }
-        TAggregationTestCase() = default;
-        TExpectedLimitChecker& MutableLimitChecker() {
-            return LimitChecker;
-        }
-        TExpectedRecordChecker& MutableRecordChecker() {
-            return RecordChecker;
-        }
-        bool GetPushdown() const {
-            return Pushdown;
-        }
-        TAggregationTestCase& SetPushdown(const bool value = true) {
-            Pushdown = value;
-            return *this;
-        }
-        bool CheckFinished() const {
-            return LimitChecker.CheckFinish();
-        }
-
-        const TString& GetQuery() const {
-            return Query;
-        }
-        TAggregationTestCase& SetQuery(const TString& value) {
-            Query = value;
-            return *this;
-        }
-        TAggregationTestCase& SetUseLlvm(const bool value) {
-            UseLlvm = value;
-            return *this;
-        }
-        const TString& GetExpectedReply() const {
-            return ExpectedReply;
-        }
-        TAggregationTestCase& SetExpectedReply(const TString& value) {
-            ExpectedReply = value;
-            return *this;
-        }
-
-        TAggregationTestCase& AddExpectedPlanOptions(const std::string& value) {
-            ExpectedPlanOptions.emplace_back(value);
-            return *this;
-        }
-
-        const std::vector<std::string>& GetExpectedPlanOptions() const {
-            return ExpectedPlanOptions;
-        }
-
-        TAggregationTestCase& SetExpectedReadNodeType(const std::string& value) {
-            ExpectedReadNodeType = value;
-            return *this;
-        }
-
-        const std::string& GetExpectedReadNodeType() const {
-            return ExpectedReadNodeType;
-        }
-    };
-
-    void TestAggregationsBase(const std::vector<TAggregationTestCase>& cases) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false)
-            .SetForceColumnTablesCompositeMarks(true);
-        TKikimrRunner kikimr(settings);
-
-        TLocalHelper(kikimr).CreateTestOlapTable();
-        auto tableClient = kikimr.GetTableClient();
-
-        {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 10000, 3000000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 11000, 3001000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 12000, 3002000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 13000, 3003000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 14000, 3004000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 20000, 2000000, 7000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 30000, 1000000, 11000);
-        }
-
-        for (auto&& i : cases) {
-            const TString queryFixed = i.GetFixedQuery();
-            {
-                auto it = tableClient.StreamExecuteScanQuery(queryFixed).GetValueSync();
-                UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-                TString result = StreamResultToYson(it);
-                if (!i.GetExpectedReply().empty()) {
-                    CompareYson(result, i.GetExpectedReply());
-                }
-            }
-            CheckPlanForAggregatePushdown(queryFixed, tableClient, i.GetExpectedPlanOptions(), i.GetExpectedReadNodeType());
-        }
-    }
-
-    void TestAggregationsInternal(const std::vector<TAggregationTestCase>& cases) {
-        TPortManager tp;
-        ui16 mbusport = tp.GetPort(2134);
-        auto settings = Tests::TServerSettings(mbusport)
-            .SetDomainName("Root")
-            .SetUseRealThreads(false)
-            .SetNodeCount(2);
-
-        Tests::TServer::TPtr server = new Tests::TServer(settings);
-
-        auto runtime = server->GetRuntime();
-        auto sender = runtime->AllocateEdgeActor();
-
-        InitRoot(server, sender);
-        Tests::NCommon::TLoggerInit(runtime).Initialize();
-
-        ui32 numShards = 1;
-        ui32 numIterations = 10;
-        TLocalHelper(*server).CreateTestOlapTable("olapTable", "olapStore", numShards, numShards);
-        const ui32 iterationPackSize = 2000;
-        for (ui64 i = 0; i < numIterations; ++i) {
-            TLocalHelper(*server).SendDataViaActorSystem("/Root/olapStore/olapTable", 0, 1000000 + i * 1000000, iterationPackSize);
-        }
-
-        TAggregationTestCase currentTest;
-        auto captureEvents = [&](TAutoPtr<IEventHandle>& ev) -> auto {
-            switch (ev->GetTypeRewrite()) {
-                case NKqp::TKqpComputeEvents::EvScanData:
-                {
-                    auto* msg = ev->Get<NKqp::TEvKqpCompute::TEvScanData>();
-                    Y_ABORT_UNLESS(currentTest.MutableLimitChecker().CheckExpectedLimitOnScanData(msg->ArrowBatch ? msg->ArrowBatch->num_rows() : 0));
-                    Y_ABORT_UNLESS(currentTest.MutableRecordChecker().CheckExpectedOnScanData(msg->ArrowBatch ? msg->ArrowBatch->num_columns() : 0));
-                    break;
-                }
-                case TEvDataShard::EvKqpScan:
-                {
-                    auto* msg = ev->Get<TEvDataShard::TEvKqpScan>();
-                    Y_ABORT_UNLESS(currentTest.MutableLimitChecker().CheckExpectedLimitOnScanTask(msg->Record.GetItemsLimit()));
-                    break;
-                }
-            }
-            return TTestActorRuntime::EEventAction::PROCESS;
-        };
-        runtime->SetObserverFunc(captureEvents);
-
-        for (auto&& i : cases) {
-            const TString queryFixed = i.GetFixedQuery();
-            currentTest = i;
-            auto streamSender = runtime->AllocateEdgeActor();
-            SendRequest(*runtime, streamSender, MakeStreamRequest(streamSender, queryFixed, false));
-            auto ev = runtime->GrabEdgeEventRethrow<NKqp::TEvKqpCompute::TEvScanData>(streamSender, TDuration::Seconds(10));
-            Y_ABORT_UNLESS(currentTest.CheckFinished());
-        }
-    }
-
-    void TestAggregations(const std::vector<TAggregationTestCase>& cases) {
-        TestAggregationsBase(cases);
-        TestAggregationsInternal(cases);
-    }
-
-    template <typename TClient>
-    auto StreamExecuteQuery(const TAggregationTestCase& testCase, TClient& client) {
-        if constexpr (std::is_same_v<NYdb::NTable::TTableClient, TClient>) {
-            return client.StreamExecuteScanQuery(testCase.GetFixedQuery()).GetValueSync();
-        } else {
-            return client.StreamExecuteQuery(
-                testCase.GetFixedQuery(),
-                NYdb::NQuery::TTxControl::BeginTx().CommitTx()).GetValueSync();
-        }
-    }
-
-    template <typename TClient>
-    void RunTestCaseWithClient(const TAggregationTestCase& testCase, TClient& client) {
-        auto it = StreamExecuteQuery(testCase, client);
-        UNIT_ASSERT_C(it.IsSuccess(), it.GetIssues().ToString());
-        TString result = StreamResultToYson(it);
-        if (!testCase.GetExpectedReply().empty()) {
-            CompareYson(result, testCase.GetExpectedReply());
-        }
-    }
-
-    void TestClickBenchBase(const std::vector<TAggregationTestCase>& cases, const bool genericQuery) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false)
-            .SetForceColumnTablesCompositeMarks(true);
-        TKikimrRunner kikimr(settings);
-
-        TClickHelper(kikimr).CreateClickBenchTable();
-        auto tableClient = kikimr.GetTableClient();
-
-
-        ui32 numIterations = 10;
-        const ui32 iterationPackSize = NSan::PlainOrUnderSanitizer(2000, 20);
-        for (ui64 i = 0; i < numIterations; ++i) {
-            WriteTestDataForClickBench(kikimr, "/Root/benchTable", 0, 1000000 + i * 1000000, iterationPackSize);
-        }
-
-        if (!genericQuery) {
-            auto tableClient = kikimr.GetTableClient();
-            for (auto&& i : cases) {
-                const TString queryFixed = i.GetFixedQuery();
-                RunTestCaseWithClient(i, tableClient);
-                CheckPlanForAggregatePushdown(queryFixed, tableClient, i.GetExpectedPlanOptions(), i.GetExpectedReadNodeType());
-            }
-        } else {
-            auto queryClient = kikimr.GetQueryClient();
-            for (auto&& i : cases) {
-                const TString queryFixed = i.GetFixedQuery();
-                RunTestCaseWithClient(i, queryClient);
-                CheckPlanForAggregatePushdown(queryFixed, queryClient, i.GetExpectedPlanOptions(), i.GetExpectedReadNodeType());
-            }
-        }
-    }
-
-    void TestClickBenchInternal(const std::vector<TAggregationTestCase>& cases) {
-        TPortManager tp;
-        ui16 mbusport = tp.GetPort(2134);
-        auto settings = Tests::TServerSettings(mbusport)
-            .SetDomainName("Root")
-            .SetUseRealThreads(false)
-            .SetNodeCount(2);
-
-        Tests::TServer::TPtr server = new Tests::TServer(settings);
-
-        auto runtime = server->GetRuntime();
-        auto sender = runtime->AllocateEdgeActor();
-
-        InitRoot(server, sender);
-
-        TClickHelper(*server).CreateClickBenchTable();
-
-        // write data
-
-        ui32 numIterations = 10;
-        const ui32 iterationPackSize = NSan::PlainOrUnderSanitizer(2000, 20);
-        for (ui64 i = 0; i < numIterations; ++i) {
-            TClickHelper(*server).SendDataViaActorSystem("/Root/benchTable", 0, 1000000 + i * 1000000,
-                                                         iterationPackSize);
-        }
-
-        TAggregationTestCase currentTest;
-        auto captureEvents = [&](TAutoPtr<IEventHandle>& ev) -> auto {
-            switch (ev->GetTypeRewrite()) {
-                case NKqp::TKqpComputeEvents::EvScanData:
-                {
-                    auto* msg = ev->Get<NKqp::TEvKqpCompute::TEvScanData>();
-                    Y_ABORT_UNLESS(currentTest.MutableLimitChecker().CheckExpectedLimitOnScanData(msg->ArrowBatch ? msg->ArrowBatch->num_rows() : 0));
-                    Y_ABORT_UNLESS(currentTest.MutableRecordChecker().CheckExpectedOnScanData(msg->ArrowBatch ? msg->ArrowBatch->num_columns() : 0));
-                    break;
-                }
-                case TEvDataShard::EvKqpScan:
-                {
-                    auto* msg = ev->Get<TEvDataShard::TEvKqpScan>();
-                    Y_ABORT_UNLESS(currentTest.MutableLimitChecker().CheckExpectedLimitOnScanTask(msg->Record.GetItemsLimit()));
-                    break;
-                }
-            }
-            return TTestActorRuntime::EEventAction::PROCESS;
-        };
-        runtime->SetObserverFunc(captureEvents);
-
-        // selects
-
-        for (auto&& i : cases) {
-            const TString queryFixed = i.GetFixedQuery();
-            currentTest = i;
-            auto streamSender = runtime->AllocateEdgeActor();
-            SendRequest(*runtime, streamSender, MakeStreamRequest(streamSender, queryFixed, false));
-            auto ev = runtime->GrabEdgeEventRethrow<NKqp::TEvKqpCompute::TEvScanData>(streamSender, TDuration::Seconds(10));
-            Y_ABORT_UNLESS(currentTest.CheckFinished());
-        }
-    }
-
-    void TestClickBench(const std::vector<TAggregationTestCase>& cases, const bool genericQuery = false) {
-        TestClickBenchBase(cases, genericQuery);
-        if (!genericQuery) {
-            TestClickBenchInternal(cases);
-        }
-    }
-
-    void TestTableWithNulls(const std::vector<TAggregationTestCase>& cases, const bool genericQuery = false) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false)
-            .SetForceColumnTablesCompositeMarks(true);
-        TKikimrRunner kikimr(settings);
-
-        Tests::NCommon::TLoggerInit(kikimr).Initialize();
-        TTableWithNullsHelper(kikimr).CreateTableWithNulls();
-        auto tableClient = kikimr.GetTableClient();
-
-        {
-            WriteTestDataForTableWithNulls(kikimr, "/Root/tableWithNulls");
-        }
-
-        if (!genericQuery) {
-            auto tableClient = kikimr.GetTableClient();
-            for (auto&& i : cases) {
-                RunTestCaseWithClient(i, tableClient);
-                CheckPlanForAggregatePushdown(i.GetFixedQuery(), tableClient,
-                    i.GetExpectedPlanOptions(), i.GetExpectedReadNodeType());
-            }
-        } else {
-            auto queryClient = kikimr.GetQueryClient();
-            for (auto&& i : cases) {
-                RunTestCaseWithClient(i, queryClient);
-                CheckPlanForAggregatePushdown(i.GetFixedQuery(), queryClient,
-                i.GetExpectedPlanOptions(), i.GetExpectedReadNodeType());
-            }
-        }
-    }
-
-    Y_UNIT_TEST(Filter_NotAllUsedFieldsInResultSet) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT id, resource_id FROM `/Root/tableWithNulls`
-                WHERE
-                    level = 5;
-            )")
-            .SetExpectedReply("[[5;#]]")
-            .AddExpectedPlanOptions("KqpOlapFilter");
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_ResultDistinctCountRI_GroupByL) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, COUNT(DISTINCT resource_id)
-                FROM `/Root/olapStore/olapTable`
-                GROUP BY level
-                ORDER BY level
-            )")
-            .SetExpectedReply("[[[0];4600u];[[1];4600u];[[2];4600u];[[3];4600u];[[4];4600u]]")
-            ;
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_ResultCountAll_FilterL) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                    SELECT
-                        COUNT(*)
-                    FROM `/Root/olapStore/olapTable`
-                    WHERE level = 2
-                )")
-            .SetExpectedReply("[[4600u;]]")
-            .AddExpectedPlanOptions("KqpOlapFilter")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg")
-            .MutableLimitChecker().SetExpectedResultCount(1)
-#else
-            .AddExpectedPlanOptions("Condense")
-#endif
-            ;
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_ResultCountL_FilterL) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    COUNT(level)
-                FROM `/Root/olapStore/olapTable`
-                WHERE level = 2
-            )")
-            .SetExpectedReply("[[4600u;]]")
-            .AddExpectedPlanOptions("KqpOlapFilter")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg")
-            .MutableLimitChecker().SetExpectedResultCount(1)
-#else
-            .AddExpectedPlanOptions("CombineCore")
-#endif
-            ;
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_ResultCountT_FilterL) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    COUNT(timestamp)
-                FROM `/Root/olapStore/olapTable`
-                WHERE level = 2
-            )")
-            .SetExpectedReply("[[4600u;]]")
-            .AddExpectedPlanOptions("KqpOlapFilter")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg")
-            .MutableLimitChecker().SetExpectedResultCount(1)
-#else
-            .AddExpectedPlanOptions("CombineCore")
-            .AddExpectedPlanOptions("KqpOlapFilter")
-#endif
-            ;
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_ResultTL_FilterL_Limit2) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    timestamp, level
-                FROM `/Root/olapStore/olapTable`
-                WHERE level = 2
-                LIMIT 2
-            )")
-            .AddExpectedPlanOptions("KqpOlapFilter")
-            .MutableLimitChecker().SetExpectedLimit(2);
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_ResultTL_FilterL_OrderT_Limit2) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    timestamp, level
-                FROM `/Root/olapStore/olapTable`
-                WHERE level = 2
-                ORDER BY timestamp
-                LIMIT 2
-            )")
-            .AddExpectedPlanOptions("KqpOlapFilter")
-            .MutableLimitChecker().SetExpectedLimit(2);
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_ResultT_FilterL_Limit2) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    timestamp
-                FROM `/Root/olapStore/olapTable`
-                WHERE level = 2
-                LIMIT 2
-            )")
-            .AddExpectedPlanOptions("KqpOlapFilter")
-            .AddExpectedPlanOptions("KqpOlapExtractMembers")
-            .MutableLimitChecker().SetExpectedLimit(2);
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_ResultT_FilterL_OrderT_Limit2) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    timestamp
-                FROM `/Root/olapStore/olapTable`
-                WHERE level = 2
-                ORDER BY timestamp
-                LIMIT 2
-            )")
-            .AddExpectedPlanOptions("KqpOlapFilter")
-            .AddExpectedPlanOptions("KqpOlapExtractMembers")
-            .MutableLimitChecker().SetExpectedLimit(2);
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_ResultL_FilterL_OrderL_Limit2) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    timestamp, level
-                FROM `/Root/olapStore/olapTable`
-                WHERE level > 1
-                ORDER BY level
-                LIMIT 2
-            )")
-            .AddExpectedPlanOptions("KqpOlapFilter");
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_ResultCountExpr) {
-        auto g = NColumnShard::TLimits::MaxBlobSizeGuard(10000);
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                    SELECT
-                        COUNT(level + 2)
-                    FROM `/Root/olapStore/olapTable`
-                )")
-            .SetExpectedReply("[[23000u;]]")
-            .AddExpectedPlanOptions("Condense1");
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Count_Null) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    COUNT(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id > 5;
-            )")
-            .SetExpectedReply("[[0u]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Count_NullMix) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    COUNT(level)
-                FROM `/Root/tableWithNulls`;
-            )")
-            .SetExpectedReply("[[5u]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Count_GroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, COUNT(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id BETWEEN 4 AND 5
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[4;1u];[5;1u]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Count_NullGroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, COUNT(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id BETWEEN 6 AND 7
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[6;0u];[7;0u]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Count_NullMixGroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, COUNT(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id > 4 AND id < 7
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[5;1u];[6;0u]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Count_GroupByNull) {
-        // Wait for KIKIMR-16940 fix
-        return;
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, COUNT(id), COUNT(level), COUNT(*)
-                FROM `/Root/tableWithNulls`
-                WHERE id > 5
-                GROUP BY level
-                ORDER BY level;
-            )")
-            .SetExpectedReply("[[#;5u;0u;5u]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Count_GroupByNullMix) {
-        // Wait for KIKIMR-16940 fix
-        return;
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, COUNT(id), COUNT(level), COUNT(*)
-                FROM `/Root/tableWithNulls`
-                WHERE id >= 5
-                GROUP BY level
-                ORDER BY level;
-            )")
-            .SetExpectedReply("[[#;5u;0u;5u];[[5];1u;1u;1u]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_NoPushdownOnDisabledEmitAggApply) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                    PRAGMA DisableEmitAggApply;
-                    SELECT
-                        COUNT(level)
-                    FROM `/Root/olapStore/olapTable`
-                )")
-            .SetExpectedReply("[[23000u;]]")
-            .AddExpectedPlanOptions("Condense1");
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(AggregationAndFilterPushdownOnDiffCols) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    COUNT(`timestamp`)
-                FROM `/Root/olapStore/olapTable`
-                WHERE level = 2
-            )")
-            .SetExpectedReply("[[4600u;]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg")
-#else
-            .AddExpectedPlanOptions("CombineCore")
-#endif
-            .AddExpectedPlanOptions("KqpOlapFilter");
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Avg) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    AVG(level), MIN(level)
-                FROM `/Root/olapStore/olapTable`
-            )")
-            .SetExpectedReply("[[[2.];[0]]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Avg_Null) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    AVG(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id > 5;
-            )")
-            .SetExpectedReply("[[#]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Avg_NullMix) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    AVG(level)
-                FROM `/Root/tableWithNulls`;
-            )")
-            .SetExpectedReply("[[[3.]]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Avg_GroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, AVG(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id BETWEEN 4 AND 5
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[4;[4.]];[5;[5.]]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Avg_NullGroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, AVG(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id BETWEEN 6 AND 7
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[6;#];[7;#]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Avg_NullMixGroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, AVG(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id > 4 AND id < 7
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[5;[5.]];[6;#]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Avg_GroupByNull) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, AVG(id), AVG(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id > 5
-                GROUP BY level
-                ORDER BY level;
-            )")
-            .SetExpectedReply("[[#;8.;#]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Avg_GroupByNullMix) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, AVG(id), AVG(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id >= 5
-                GROUP BY level
-                ORDER BY level;
-            )")
-            .SetExpectedReply("[[#;8.;#];[[5];5.;[5.]]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Sum) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    SUM(level)
-                FROM `/Root/olapStore/olapTable`
-            )")
-            .SetExpectedReply("[[[46000;]]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Sum_Null) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    SUM(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id > 5;
-            )")
-            .SetExpectedReply("[[#]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Sum_NullMix) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    SUM(level)
-                FROM `/Root/tableWithNulls`;
-            )")
-            .SetExpectedReply("[[[15]]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Sum_GroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, SUM(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id BETWEEN 4 AND 5
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[4;[4]];[5;[5]]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Sum_NullGroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, SUM(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id BETWEEN 6 AND 7
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[6;#];[7;#]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Sum_NullMixGroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, SUM(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id > 4 AND id < 7
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[5;[5]];[6;#]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Sum_GroupByNull) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, SUM(id), SUM(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id > 5
-                GROUP BY level
-                ORDER BY level;
-            )")
-            .SetExpectedReply("[[#;40;#]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Sum_GroupByNullMix) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, SUM(id), SUM(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id >= 5
-                GROUP BY level
-                ORDER BY level;
-            )")
-            .SetExpectedReply("[[#;40;#];[[5];5;[5]]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_SumL_GroupL_OrderL) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, SUM(level)
-                FROM `/Root/olapStore/olapTable`
-                GROUP BY level
-                ORDER BY level
-            )")
-            .SetExpectedReply("[[[0];[0]];[[1];[4600]];[[2];[9200]];[[3];[13800]];[[4];[18400]]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_MinL) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    MIN(level)
-                FROM `/Root/olapStore/olapTable`
-            )")
-            .SetExpectedReply("[[[0]]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_MaxL) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    MAX(level)
-                FROM `/Root/olapStore/olapTable`
-            )")
-            .SetExpectedReply("[[[4]]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_MinR_GroupL_OrderL) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, MIN(resource_id)
-                FROM `/Root/olapStore/olapTable`
-                GROUP BY level
-                ORDER BY level
-            )")
-            .SetExpectedReply("[[[0];[\"10000\"]];[[1];[\"10001\"]];[[2];[\"10002\"]];[[3];[\"10003\"]];[[4];[\"10004\"]]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_MaxR_GroupL_OrderL) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, MAX(resource_id)
-                FROM `/Root/olapStore/olapTable`
-                GROUP BY level
-                ORDER BY level
-            )")
-            .SetExpectedReply("[[[0];[\"40995\"]];[[1];[\"40996\"]];[[2];[\"40997\"]];[[3];[\"40998\"]];[[4];[\"40999\"]]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_ProjectionOrder) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    resource_id, level, count(*) as c
-                FROM `/Root/olapStore/olapTable`
-                GROUP BY resource_id, level
-                ORDER BY c, resource_id DESC LIMIT 3
-            )")
-            .SetExpectedReply("[[[\"40999\"];[4];1u];[[\"40998\"];[3];1u];[[\"40997\"];[2];1u]]")
-            .SetExpectedReadNodeType("Aggregate-TableFullScan");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-        TestAggregations({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Some) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT SOME(level) FROM `/Root/tableWithNulls` WHERE id=1
-            )")
-            .SetExpectedReply("[[[1]]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Some_Null) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT SOME(level) FROM `/Root/tableWithNulls` WHERE id > 5
-            )")
-            .SetExpectedReply("[[#]]")
-#if SSA_RUNTIME_VERSION >= 2U
-            .AddExpectedPlanOptions("TKqpOlapAgg");
-#else
-            .AddExpectedPlanOptions("CombineCore");
-#endif
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Some_GroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, SOME(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id BETWEEN 4 AND 5
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[4;[4]];[5;[5]]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Some_NullGroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, SOME(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id BETWEEN 6 AND 7
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[6;#];[7;#]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Some_NullMixGroupBy) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, SOME(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id > 4 AND id < 7
-                GROUP BY id
-                ORDER BY id;
-            )")
-            .SetExpectedReply("[[5;[5]];[6;#]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Some_GroupByNullMix) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, SOME(id), SOME(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id BETWEEN 5 AND 6
-                GROUP BY level
-                ORDER BY level;
-            )")
-            .SetExpectedReply("[[#;6;#];[[5];5;[5]]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Aggregation_Some_GroupByNull) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, SOME(id), SOME(level)
-                FROM `/Root/tableWithNulls`
-                WHERE id = 6
-                GROUP BY level
-                ORDER BY level;
-            )")
-            .SetExpectedReply("[[#;6;#]]");
-        testCase.FillExpectedAggregationGroupByPlanOptions();
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(ClickBenchSmoke) {
-        TAggregationTestCase q7;
-        q7.SetQuery(R"(
-                SELECT
-                    AdvEngineID, COUNT(*) as c
-                FROM `/Root/benchTable`
-                WHERE AdvEngineID != 0
-                GROUP BY AdvEngineID
-                ORDER BY c DESC
-            )")
-            //.SetExpectedReply("[[[\"40999\"];[4];1u];[[\"40998\"];[3];1u];[[\"40997\"];[2];1u]]")
-            // Should be fixed in https://st.yandex-team.ru/KIKIMR-17009
-            // .SetExpectedReadNodeType("TableFullScan");
-            .SetExpectedReadNodeType("Aggregate-TableFullScan");
-        q7.FillExpectedAggregationGroupByPlanOptions();
-
-        TAggregationTestCase q9;
-        q9.SetQuery(R"(
-                SELECT
-                    RegionID, SUM(AdvEngineID), COUNT(*) AS c, avg(ResolutionWidth), COUNT(DISTINCT UserID)
-                FROM `/Root/benchTable`
-                GROUP BY RegionID
-                ORDER BY c DESC
-                LIMIT 10
-            )")
-            //.SetExpectedReply("[[[\"40999\"];[4];1u];[[\"40998\"];[3];1u];[[\"40997\"];[2];1u]]")
-            // Should be fixed in https://st.yandex-team.ru/KIKIMR-17009
-            // .SetExpectedReadNodeType("TableFullScan");
-            .SetExpectedReadNodeType("Aggregate-TableFullScan");
-        q9.FillExpectedAggregationGroupByPlanOptions();
-
-        TAggregationTestCase q12;
-        q12.SetQuery(R"(
-                SELECT
-                    SearchPhrase, count(*) AS c
-                FROM `/Root/benchTable`
-                WHERE SearchPhrase != ''
-                GROUP BY SearchPhrase
-                ORDER BY c DESC
-                LIMIT 10;
-            )")
-            //.SetExpectedReply("[[[\"40999\"];[4];1u];[[\"40998\"];[3];1u];[[\"40997\"];[2];1u]]")
-            // Should be fixed in https://st.yandex-team.ru/KIKIMR-17009
-            // .SetExpectedReadNodeType("TableFullScan");
-            .SetExpectedReadNodeType("Aggregate-TableFullScan");
-        q12.FillExpectedAggregationGroupByPlanOptions();
-
-        TAggregationTestCase q14;
-        q14.SetQuery(R"(
-                SELECT
-                    SearchEngineID, SearchPhrase, count(*) AS c
-                FROM `/Root/benchTable`
-                WHERE SearchPhrase != ''
-                GROUP BY SearchEngineID, SearchPhrase
-                ORDER BY c DESC
-                LIMIT 10;
-            )")
-            //.SetExpectedReply("[[[\"40999\"];[4];1u];[[\"40998\"];[3];1u];[[\"40997\"];[2];1u]]")
-            // Should be fixed in https://st.yandex-team.ru/KIKIMR-17009
-            // .SetExpectedReadNodeType("TableFullScan");
-            .SetExpectedReadNodeType("Aggregate-TableFullScan");
-        q14.FillExpectedAggregationGroupByPlanOptions();
-
-        TAggregationTestCase q22;
-        q22.SetQuery(R"(
-                SELECT
-                    SearchPhrase, MIN(URL), MIN(Title), COUNT(*) AS c, COUNT(DISTINCT UserID)
-                FROM `/Root/benchTable`
-                WHERE Title LIKE '%Google%' AND URL NOT LIKE '%.google.%' AND SearchPhrase <> ''
-                GROUP BY SearchPhrase
-                ORDER BY c DESC
-                LIMIT 10;
-            )")
-            .AddExpectedPlanOptions("KqpOlapFilter")
-            .SetExpectedReadNodeType("Aggregate-TableFullScan");
-        q22.FillExpectedAggregationGroupByPlanOptions();
-
-        TAggregationTestCase q39;
-        q39.SetQuery(R"(
-                SELECT TraficSourceID, SearchEngineID, AdvEngineID, Src, Dst, COUNT(*) AS PageViews
-                FROM `/Root/benchTable`
-                WHERE CounterID = 62 AND EventDate >= Date('2013-07-01') AND EventDate <= Date('2013-07-31') AND IsRefresh == 0
-                GROUP BY
-                    TraficSourceID, SearchEngineID, AdvEngineID, IF (SearchEngineID = 0 AND AdvEngineID = 0, Referer, '') AS Src,
-                    URL AS Dst
-                ORDER BY PageViews DESC
-                LIMIT 10;
-            )")
-            .AddExpectedPlanOptions("KqpOlapFilter")
-            .SetExpectedReadNodeType("Aggregate-Filter-TableFullScan");
-        q39.FillExpectedAggregationGroupByPlanOptions();
-
-        std::vector<TAggregationTestCase> cases = {q7, q9, q12, q14, q22, q39};
-        for (auto&& c : cases) {
-            c.SetUseLlvm(NSan::PlainOrUnderSanitizer(true, false));
-        }
-
-        TestClickBench(cases);
-    }
-
-    Y_UNIT_TEST(StatsSysView) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        static ui32 numKinds = 2;
-
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        TLocalHelper(kikimr).CreateTestOlapTable();
-        for (ui64 i = 0; i < 100; ++i) {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 0, 1000000 + i * 10000, 1000);
-        }
-
-        auto tableClient = kikimr.GetTableClient();
-        auto selectQuery = TString(R"(
-            SELECT PathId, Kind, TabletId, Sum(Rows) as Rows
-            FROM `/Root/olapStore/.sys/store_primary_index_stats`
-            GROUP BY PathId, Kind, TabletId
-            ORDER BY TabletId, Kind, PathId
-        )");
-
-        auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-        UNIT_ASSERT_VALUES_EQUAL(rows.size(), numKinds*3);
-        UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("PathId")), 3ull);
-        UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[0].at("Kind")), "INSERTED");
-        UNIT_ASSERT_GE(GetUint64(rows[0].at("TabletId")), 72075186224037888ull);
-        UNIT_ASSERT_GE(GetUint64(rows[2].at("TabletId")), 72075186224037889ull);
-        UNIT_ASSERT_GE(GetUint64(rows[4].at("TabletId")), 72075186224037890ull);
-        UNIT_ASSERT_GE(GetUint64(rows[1].at("TabletId")), GetUint64(rows[0].at("TabletId")));
-        UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[2].at("Kind")), "INSERTED");
-        UNIT_ASSERT_GE(GetUint64(rows[2].at("TabletId")), GetUint64(rows[1].at("TabletId")));
-        UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[3].at("PathId")), 3ull);
-        UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[3].at("Kind")), "SPLIT_COMPACTED");
-        UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[3].at("TabletId")), GetUint64(rows[2].at("TabletId")));
-        UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[4].at("Kind")), "INSERTED");
-        UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[4].at("TabletId")), GetUint64(rows[5].at("TabletId")));
-        UNIT_ASSERT_GE(
-            GetUint64(rows[0].at("Rows")) + GetUint64(rows[1].at("Rows")) + GetUint64(rows[2].at("Rows")) +
-            GetUint64(rows[3].at("Rows")) + GetUint64(rows[4].at("Rows")) + GetUint64(rows[5].at("Rows")),
-            0.3*0.9*100*1000); // >= 90% of 100K inserted rows
-    }
-
-    Y_UNIT_TEST(StatsSysViewTable) {
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        static ui32 numKinds = 5;
-
-        TLocalHelper(kikimr).CreateTestOlapTable("olapTable_1");
-        TLocalHelper(kikimr).CreateTestOlapTable("olapTable_2");
-        for (ui64 i = 0; i < 10; ++i) {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable_1", 0, 1000000 + i*10000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable_2", 0, 1000000 + i*10000, 2000);
-        }
-
-        auto tableClient = kikimr.GetTableClient();
-        {
-            auto selectQuery = TString(R"(
-                SELECT PathId, Kind, TabletId
-                FROM `/Root/olapStore/olapTable_1/.sys/primary_index_stats`
-                GROUP BY PathId, TabletId, Kind
-                ORDER BY PathId, TabletId, Kind
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            UNIT_ASSERT_GT(rows.size(), 1*numKinds);
-            UNIT_ASSERT_LE(rows.size(), 3*numKinds);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows.front().at("PathId")), 3ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows.back().at("PathId")), 3ull);
-        }
-        {
-            auto selectQuery = TString(R"(
-                SELECT PathId, Kind, TabletId
-                FROM `/Root/olapStore/olapTable_2/.sys/primary_index_stats`
-                GROUP BY PathId, TabletId, Kind
-                ORDER BY PathId, TabletId, Kind
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            UNIT_ASSERT_GT(rows.size(), 1*numKinds);
-            UNIT_ASSERT_LE(rows.size(), 3*numKinds);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows.front().at("PathId")), 4ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows.back().at("PathId")), 4ull);
-        }
-        {
-            auto selectQuery = TString(R"(
-                SELECT *
-                FROM `/Root/olapStore/olapTable_1/.sys/primary_index_stats`
-                WHERE
-                    PathId > UInt64("3")
-                ORDER BY PathId, Kind, TabletId
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 0);
-        }
-    }
-
-    Y_UNIT_TEST(StatsSysViewEnumStringBytes) {
-        ui64 rawBytesPK1;
-        ui64 bytesPK1;
-        {
-            auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-            csController->SetPeriodicWakeupActivationPeriod(TDuration::Seconds(1));
-            auto settings = TKikimrSettings()
-                .SetWithSampleTables(false);
-            TKikimrRunner kikimr(settings);
-            Tests::NCommon::TLoggerInit(kikimr).Initialize();
-            TTypedLocalHelper helper("", kikimr, "olapTable", "olapStore12");
-            helper.CreateTestOlapTable();
-            helper.FillPKOnly(0, 800000);
-            helper.GetVolumes(rawBytesPK1, bytesPK1, false);
-        }
-
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        ui64 rawBytesUnpack1PK = 0;
-        ui64 bytesUnpack1PK = 0;
-        ui64 rawBytesPackAndUnpack2PK;
-        ui64 bytesPackAndUnpack2PK;
-        const ui32 rowsCount = 800000;
-        const ui32 groupsCount = 512;
-        {
-            auto settings = TKikimrSettings()
-                .SetWithSampleTables(false);
-            TKikimrRunner kikimr(settings);
-            Tests::NCommon::TLoggerInit(kikimr).Initialize();
-            TTypedLocalHelper helper("Utf8", kikimr);
-            helper.CreateTestOlapTable();
-            NArrow::NConstruction::TStringPoolFiller sPool(groupsCount, 52);
-            helper.FillTable(sPool, 0, rowsCount);
-            helper.PrintCount();
-            {
-                auto d = helper.GetDistribution();
-                Y_ABORT_UNLESS(d.GetCount() == rowsCount);
-                Y_ABORT_UNLESS(d.GetGroupsCount() == groupsCount);
-                Y_ABORT_UNLESS(d.GetMaxCount() - d.GetMinCount() <= 1);
-            }
-            helper.GetVolumes(rawBytesUnpack1PK, bytesUnpack1PK, false);
-            Sleep(TDuration::Seconds(5));
-            auto tableClient = kikimr.GetTableClient();
-            helper.ExecuteSchemeQuery(TStringBuilder() << "ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=ALTER_COLUMN, NAME=field, `ENCODING.DICTIONARY.ENABLED`=`true`);");
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=ALTER_COLUMN, NAME=field1, `ENCODING.DICTIONARY.ENABLED`=`true`);", EStatus::SCHEME_ERROR);
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=ALTER_COLUMN, NAME=field, `ENCODING.DICTIONARY.ENABLED1`=`true`);", EStatus::GENERIC_ERROR);
-            Sleep(TDuration::Seconds(5));
-            helper.FillTable(sPool, 1, rowsCount);
-            Sleep(TDuration::Seconds(5));
-            {
-                helper.GetVolumes(rawBytesPackAndUnpack2PK, bytesPackAndUnpack2PK, false);
-                helper.PrintCount();
-                {
-                    auto d = helper.GetDistribution();
-                    Cerr << d.DebugString() << Endl;
-                    Y_ABORT_UNLESS(d.GetCount() == 2 * rowsCount);
-                    Y_ABORT_UNLESS(d.GetGroupsCount() == groupsCount);
-                    Y_ABORT_UNLESS(d.GetMaxCount() - d.GetMinCount() <= 2);
-                }
-            }
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=ALTER_COLUMN, NAME=field, `SERIALIZER.CLASS_NAME`=`ARROW_SERIALIZER`, `COMPRESSION.TYPE`=`zstd`);");
-        }
-        const ui64 rawBytesUnpack = rawBytesUnpack1PK - rawBytesPK1;
-        const ui64 bytesUnpack = bytesUnpack1PK - bytesPK1;
-        const ui64 rawBytesPack = rawBytesPackAndUnpack2PK - rawBytesUnpack1PK - rawBytesPK1;
-        const ui64 bytesPack = bytesPackAndUnpack2PK - bytesUnpack1PK - bytesPK1;
-        TStringBuilder result;
-        result << "unpacked data: " << rawBytesUnpack << " / " << bytesUnpack << Endl;
-        result << "packed data: " << rawBytesPack << " / " << bytesPack << Endl;
-        result << "frq_diff: " << 1.0 * bytesPack / bytesUnpack << Endl;
-        result << "frq_compression: " << 1.0 * bytesPack / rawBytesPack << Endl;
-        result << "pk_size : " << rawBytesPK1 << " / " << bytesPK1 << Endl;
-        Cerr << result << Endl;
-        Y_ABORT_UNLESS(bytesPack / bytesUnpack < 0.1);
-    }
-
-    Y_UNIT_TEST(StatsSysViewBytesPackActualization) {
-        ui64 rawBytesPK1;
-        ui64 bytesPK1;
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        csController->SetPeriodicWakeupActivationPeriod(TDuration::Seconds(1));
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        Tests::NCommon::TLoggerInit(kikimr).Initialize();
-        TTypedLocalHelper helper("", kikimr, "olapTable", "olapStore");
-        helper.CreateTestOlapTable();
-        helper.FillPKOnly(0, 800000);
-        helper.GetVolumes(rawBytesPK1, bytesPK1, false, {"pk_int"});
-        auto tableClient = kikimr.GetTableClient();
-        {
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=ALTER_COLUMN, NAME=pk_int, `SERIALIZER.CLASS_NAME`=`ARROW_SERIALIZER`, `COMPRESSION.TYPE`=`zstd`);");
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_OPTIONS, SCHEME_NEED_ACTUALIZATION=`true`);");
-            csController->WaitActualization(TDuration::Seconds(10));
-            ui64 rawBytesPK2;
-            ui64 bytesPK2;
-            helper.GetVolumes(rawBytesPK2, bytesPK2, false, {"pk_int"});
-            AFL_VERIFY(rawBytesPK2 == rawBytesPK1)("pk1", rawBytesPK1)("pk2", rawBytesPK2);
-            AFL_VERIFY(bytesPK2 < bytesPK1 / 3)("pk1", bytesPK1)("pk2", bytesPK2);
-        }
-        {
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=ALTER_COLUMN, NAME=pk_int, `SERIALIZER.CLASS_NAME`=`ARROW_SERIALIZER`, `COMPRESSION.TYPE`=`lz4`);");
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_OPTIONS, SCHEME_NEED_ACTUALIZATION=`true`);");
-            csController->WaitActualization(TDuration::Seconds(10));
-            ui64 rawBytesPK2;
-            ui64 bytesPK2;
-            helper.GetVolumes(rawBytesPK2, bytesPK2, false, {"pk_int"});
-            AFL_VERIFY(rawBytesPK2 == rawBytesPK1)("pk1", rawBytesPK1)("pk2", rawBytesPK2);
-            AFL_VERIFY(bytesPK2 < bytesPK1 * 1.01 && bytesPK1 < bytesPK2 * 1.01)("pk1", bytesPK1)("pk2", bytesPK2);
-        }
-    }
-
-    Y_UNIT_TEST(StatsSysViewBytesColumnActualization) {
-        ui64 rawBytes1;
-        ui64 bytes1;
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        csController->SetPeriodicWakeupActivationPeriod(TDuration::Seconds(1));
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        Tests::NCommon::TLoggerInit(kikimr).Initialize();
-        TTypedLocalHelper helper("Utf8", kikimr);
-        helper.CreateTestOlapTable();
-        NArrow::NConstruction::TStringPoolFiller sPool(3, 52);
-        helper.FillTable(sPool, 0, 800000);
-        helper.GetVolumes(rawBytes1, bytes1, false, {"new_column_ui64"});
-        AFL_VERIFY(rawBytes1 == 0);
-        AFL_VERIFY(bytes1 == 0);
-        auto tableClient = kikimr.GetTableClient();
-        {
-            helper.ExecuteSchemeQuery("ALTER TABLESTORE `/Root/olapStore` ADD COLUMN new_column_ui64 Uint64;");
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_OPTIONS, SCHEME_NEED_ACTUALIZATION=`true`);");
-            csController->WaitActualization(TDuration::Seconds(10));
-            ui64 rawBytes2;
-            ui64 bytes2;
-            helper.GetVolumes(rawBytes2, bytes2, false, {"new_column_ui64"});
-            AFL_VERIFY(rawBytes2 == 6500041)("real", rawBytes2);
-            AFL_VERIFY(bytes2 == 45360)("b", bytes2);
-        }
-    }
-
-    Y_UNIT_TEST(StatsSysViewBytesDictActualization) {
-        ui64 rawBytes1;
-        ui64 bytes1;
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        csController->SetPeriodicWakeupActivationPeriod(TDuration::Seconds(1));
-        auto settings = TKikimrSettings()
-            .SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        Tests::NCommon::TLoggerInit(kikimr).Initialize();
-        TTypedLocalHelper helper("Utf8", kikimr);
-        helper.CreateTestOlapTable();
-        NArrow::NConstruction::TStringPoolFiller sPool(3, 52);
-        helper.FillTable(sPool, 0, 800000);
-        helper.GetVolumes(rawBytes1, bytes1, false, {"field"});
-        auto tableClient = kikimr.GetTableClient();
-        {
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=ALTER_COLUMN, NAME=field, `ENCODING.DICTIONARY.ENABLED`=`true`);");
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_OPTIONS, SCHEME_NEED_ACTUALIZATION=`true`);");
-            csController->WaitActualization(TDuration::Seconds(10));
-            ui64 rawBytes2;
-            ui64 bytes2;
-            helper.GetVolumes(rawBytes2, bytes2, false, {"field"});
-            AFL_VERIFY(rawBytes2 == rawBytes1)("f1", rawBytes1)("f2", rawBytes2);
-            AFL_VERIFY(bytes2 < bytes1 * 0.5)("f1", bytes1)("f2", bytes2);
-        }
-        {
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=ALTER_COLUMN, NAME=field, `ENCODING.DICTIONARY.ENABLED`=`false`);");
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_OPTIONS, SCHEME_NEED_ACTUALIZATION=`true`);");
-            csController->WaitActualization(TDuration::Seconds(10));
-            ui64 rawBytes2;
-            ui64 bytes2;
-            helper.GetVolumes(rawBytes2, bytes2, false, {"field"});
-            AFL_VERIFY(rawBytes2 == rawBytes1)("f1", rawBytes1)("f2", rawBytes2);
-            AFL_VERIFY(bytes2 < bytes1 * 1.01 && bytes1 < bytes2 * 1.01)("f1", bytes1)("f2", bytes2);
-        }
-    }
-
-    Y_UNIT_TEST(StatsSysViewBytesDictStatActualization) {
-        ui64 rawBytes1;
-        ui64 bytes1;
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        csController->SetPeriodicWakeupActivationPeriod(TDuration::Seconds(1));
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        Tests::NCommon::TLoggerInit(kikimr).Initialize();
-        TTypedLocalHelper helper("Utf8", kikimr);
-        helper.CreateTestOlapTable();
-        NArrow::NConstruction::TStringPoolFiller sPool(3, 52);
-        helper.FillTable(sPool, 0, 800000);
-        helper.GetVolumes(rawBytes1, bytes1, false, {"field"});
-        auto tableClient = kikimr.GetTableClient();
-        {
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=ALTER_COLUMN, NAME=field, `ENCODING.DICTIONARY.ENABLED`=`true`);");
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_STAT, NAME=field_var, TYPE=variability, FEATURES=`{\"column_name\" : \"field\"}`);");
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_STAT, NAME=pk_int_max, TYPE=max, FEATURES=`{\"column_name\" : \"pk_int\"}`);");
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_OPTIONS, SCHEME_NEED_ACTUALIZATION=`true`);");
-            csController->WaitActualization(TDuration::Seconds(10));
-            ui64 rawBytes2;
-            ui64 bytes2;
-            helper.GetVolumes(rawBytes2, bytes2, false, {"field"});
-            AFL_VERIFY(rawBytes2 == rawBytes1)("f1", rawBytes1)("f2", rawBytes2);
-            AFL_VERIFY(bytes2 < bytes1 * 0.5)("f1", bytes1)("f2", bytes2);
-            std::vector<NKikimrColumnShardStatisticsProto::TPortionStorage> stats;
-            helper.GetStats(stats, true);
-            for (auto&& i : stats) {
-                AFL_VERIFY(i.ScalarsSize() == 2);
-                AFL_VERIFY(i.GetScalars()[0].GetUint32() == 3);
-            }
-        }
-        {
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=DROP_STAT, NAME=pk_int_max);");
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_OPTIONS, SCHEME_NEED_ACTUALIZATION=`true`);");
-            csController->WaitActualization(TDuration::Seconds(10));
-            std::vector<NKikimrColumnShardStatisticsProto::TPortionStorage> stats;
-            helper.GetStats(stats, true);
-            for (auto&& i : stats) {
-                AFL_VERIFY(i.ScalarsSize() == 1);
-                AFL_VERIFY(i.GetScalars()[0].GetUint32() == 3);
-            }
-        }
-        {
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_STAT, NAME=pk_int_max, TYPE=max, FEATURES=`{\"column_name\" : \"pk_int\"}`);");
-            helper.ExecuteSchemeQuery("ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_OPTIONS, SCHEME_NEED_ACTUALIZATION=`true`);");
-            csController->WaitActualization(TDuration::Seconds(10));
-            std::vector<NKikimrColumnShardStatisticsProto::TPortionStorage> stats;
-            helper.GetStats(stats, true);
-            for (auto&& i : stats) {
-                AFL_VERIFY(i.ScalarsSize() == 2);
-                AFL_VERIFY(i.GetScalars()[0].GetUint32() == 3);
-            }
-        }
-    }
-
-    Y_UNIT_TEST(StatsUsage) {
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        {
-            auto settings = TKikimrSettings().SetWithSampleTables(false);
-            TKikimrRunner kikimr(settings);
-            Tests::NCommon::TLoggerInit(kikimr).Initialize();
-            TTypedLocalHelper helper("Utf8", kikimr);
-            helper.CreateTestOlapTable();
-            auto tableClient = kikimr.GetTableClient();
-            {
-                auto alterQuery = TStringBuilder() << "ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_STAT, TYPE=max, NAME=max_pk_int, FEATURES=`{\"column_name\": \"pk_int\"}`);";
-                auto session = tableClient.CreateSession().GetValueSync().GetSession();
-                auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-                UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-            }
-            {
-                auto alterQuery = TStringBuilder() << "ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_STAT, TYPE=max, NAME=max_field, FEATURES=`{\"column_name\": \"field\"}`);";
-                auto session = tableClient.CreateSession().GetValueSync().GetSession();
-                auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-                UNIT_ASSERT_VALUES_UNEQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-            }
-            {
-                auto alterQuery = TStringBuilder() << "ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_STAT, TYPE=max, NAME=max_pk_int, FEATURES=`{\"column_name\": \"pk_int\"}`);";
-                auto session = tableClient.CreateSession().GetValueSync().GetSession();
-                auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-                UNIT_ASSERT_VALUES_UNEQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-            }
-            {
-                auto alterQuery = TStringBuilder() << "ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=DROP_STAT, NAME=max_pk_int);";
-                auto session = tableClient.CreateSession().GetValueSync().GetSession();
-                auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-                UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-            }
-        }
-    }
-
-    Y_UNIT_TEST(StatsUsageWithTTL) {
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        {
-            auto settings = TKikimrSettings().SetWithSampleTables(false);
-            TKikimrRunner kikimr(settings);
-            Tests::NCommon::TLoggerInit(kikimr).Initialize();
-            TTypedLocalHelper helper("Utf8", kikimr);
-            helper.CreateTestOlapTable();
-            auto tableClient = kikimr.GetTableClient();
-            {
-                auto alterQuery = TStringBuilder() << "ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=UPSERT_STAT, TYPE=max, NAME=max_ts, FEATURES=`{\"column_name\": \"ts\"}`);";
-                auto session = tableClient.CreateSession().GetValueSync().GetSession();
-                auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-                UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-            }
-            {
-                auto alterQuery = TStringBuilder() << "ALTER TABLE `/Root/olapStore/olapTable` SET (TTL = Interval(\"P1D\") ON ts);";
-                auto session = tableClient.CreateSession().GetValueSync().GetSession();
-                auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-                UNIT_ASSERT_VALUES_EQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-            }
-            {
-                auto alterQuery = TStringBuilder() << "ALTER OBJECT `/Root/olapStore` (TYPE TABLESTORE) SET (ACTION=DROP_STAT, NAME=max_ts);";
-                auto session = tableClient.CreateSession().GetValueSync().GetSession();
-                auto alterResult = session.ExecuteSchemeQuery(alterQuery).GetValueSync();
-                UNIT_ASSERT_VALUES_UNEQUAL_C(alterResult.GetStatus(), EStatus::SUCCESS, alterResult.GetIssues().ToString());
-            }
-        }
-    }
-
-    namespace {
-    class TTransferStatus {
-    private:
-        YDB_ACCESSOR(bool, Proposed, false);
-        YDB_ACCESSOR(bool, Confirmed, false);
-        YDB_ACCESSOR(bool, Finished, false);
-    public:
-        void Reset() {
-            Confirmed = false;
-            Proposed = false;
-            Finished = false;
-        }
-    };
-
-    static TMutex CSTransferStatusesMutex;
-    static std::shared_ptr<TTransferStatus> CSTransferStatus = std::make_shared<TTransferStatus>();
-    }
-
-    class TTestController: public NOlap::NDataSharing::IInitiatorController {
-    private:
-        static const inline auto Registrator = TFactory::TRegistrator<TTestController>("test");
-    protected:
-        virtual void DoProposeError(const TString& sessionId, const TString& message) const override {
-            AFL_VERIFY(false)("session_id", sessionId)("message", message);
-        }
-        virtual void DoProposeSuccess(const TString& sessionId) const override {
-            CSTransferStatus->SetProposed(true);
-            AFL_NOTICE(NKikimrServices::TX_COLUMNSHARD)("event", "sharing_proposed")("session_id", sessionId);
-        }
-        virtual void DoConfirmSuccess(const TString& sessionId) const override {
-            CSTransferStatus->SetConfirmed(true);
-            AFL_NOTICE(NKikimrServices::TX_COLUMNSHARD)("event", "sharing_confirmed")("session_id", sessionId);
-        }
-        virtual void DoFinished(const TString& sessionId) const override {
-            CSTransferStatus->SetFinished(true);
-            AFL_NOTICE(NKikimrServices::TX_COLUMNSHARD)("event", "sharing_finished")("session_id", sessionId);
-        }
-        virtual void DoStatus(const NOlap::NDataSharing::TStatusContainer& status) const override {
-            AFL_NOTICE(NKikimrServices::TX_COLUMNSHARD)("event", "status")("info", status.SerializeToProto().DebugString());
-        }
-        virtual TConclusionStatus DoDeserializeFromProto(const NKikimrColumnShardDataSharingProto::TInitiator::TController& /*proto*/) override {
-            return TConclusionStatus::Success();
-        }
-        virtual void DoSerializeToProto(NKikimrColumnShardDataSharingProto::TInitiator::TController& /*proto*/) const override {
-            
-        }
-
-        virtual TString GetClassName() const override {
-            return "test";
-        }
-    };
-
-    class TSharingDataTestCase {
-    private:
-        const ui32 ShardsCount;
-        TKikimrRunner& Kikimr;
-        TTypedLocalHelper Helper;
-        NYDBTest::TControllers::TGuard<NYDBTest::NColumnShard::TController> Controller;
-        std::vector<ui64> ShardIds;
-        std::vector<ui64> PathIds;
-        YDB_ACCESSOR(bool, RebootTablet, false);
-    public:
-        const TTypedLocalHelper& GetHelper() const {
-            return Helper;
-        }
-
-        void AddRecords(const ui32 recordsCount, const double kff = 0) {
-            Helper.FillPKOnly(kff, recordsCount);
-        }
-
-        TSharingDataTestCase(const ui32 shardsCount, TKikimrRunner& kikimr)
-            : ShardsCount(shardsCount)
-            , Kikimr(kikimr)
-            , Helper("", Kikimr, "olapTable", "olapStore12")
-            , Controller(NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>())
-        {
-            Controller->SetCompactionControl(NYDBTest::EOptimizerCompactionWeightControl::Disable);
-            Controller->SetExpectedShardsCount(ShardsCount);
-            Controller->SetPeriodicWakeupActivationPeriod(TDuration::Seconds(1));
-            Controller->SetReadTimeoutClean(TDuration::Seconds(1));
-
-            Tests::NCommon::TLoggerInit(Kikimr).SetComponents({NKikimrServices::TX_COLUMNSHARD}, "CS").Initialize();
-
-            Helper.CreateTestOlapTable(ShardsCount, ShardsCount);
-            ShardIds = Controller->GetShardActualIds();
-            AFL_VERIFY(ShardIds.size() == ShardsCount)("count", ShardIds.size())("ids", JoinSeq(",", ShardIds));
-            std::set<ui64> pathIdsSet;
-            for (auto&& i : ShardIds) {
-                auto pathIds = Controller->GetPathIds(i);
-                pathIdsSet.insert(pathIds.begin(), pathIds.end());
-            }
-            PathIds = std::vector<ui64>(pathIdsSet.begin(), pathIdsSet.end());
-            AFL_VERIFY(PathIds.size() == 1)("count", PathIds.size())("ids", JoinSeq(",", PathIds));
-        }
-
-        void WaitNormalization() {
-            Controller->SetCompactionControl(NYDBTest::EOptimizerCompactionWeightControl::Force);
-            const auto start = TInstant::Now();
-            while (!Controller->IsTrivialLinks() && TInstant::Now() - start < TDuration::Seconds(30)) {
-                Cerr << "WAIT_TRIVIAL_LINKS..." << Endl;
-                Sleep(TDuration::Seconds(1));
-            }
-            AFL_VERIFY(Controller->IsTrivialLinks());
-            Controller->CheckInvariants();
-        }
-
-        void Execute(const ui64 destinationIdx, const std::vector<ui64>& sourceIdxs, const bool move, const NOlap::TSnapshot& snapshot, const std::set<ui64>& pathIdxs) {
-            AFL_VERIFY(destinationIdx < ShardIds.size());
-            const ui64 destination = ShardIds[destinationIdx];
-            std::vector<ui64> sources;
-            for (auto&& i : sourceIdxs) {
-                AFL_VERIFY(i < ShardIds.size());
-                sources.emplace_back(ShardIds[i]);
-            }
-            std::set<ui64> pathIds;
-            for (auto&& i : pathIdxs) {
-                AFL_VERIFY(i < PathIds.size());
-                AFL_VERIFY(pathIds.emplace(PathIds[i]).second);
-            }
-            Cerr << "SHARING: " << JoinSeq(",", sources) << "->" << destination << Endl;
-            THashMap<ui64, ui64> pathIdsRemap;
-            for (auto&& i : pathIds) {
-                pathIdsRemap.emplace(i, i);
-            }
-            THashSet<NOlap::TTabletId> sourceTablets;
-            for (auto&& i : sources) {
-                AFL_VERIFY(sourceTablets.emplace((NOlap::TTabletId)i).second);
-            }
-            const TString sessionId = TGUID::CreateTimebased().AsUuidString();
-            NOlap::NDataSharing::TTransferContext transferContext((NOlap::TTabletId)destination, sourceTablets, snapshot, move);
-            NOlap::NDataSharing::TDestinationSession session(std::make_shared<TTestController>(), pathIdsRemap, sessionId, transferContext);
-            Kikimr.GetTestServer().GetRuntime()->Send(MakePipePeNodeCacheID(false), NActors::TActorId(), new TEvPipeCache::TEvForward(
-                new NOlap::NDataSharing::NEvents::TEvProposeFromInitiator(session), destination, false));
-            {
-                const TInstant start = TInstant::Now();
-                while (!CSTransferStatus->GetProposed() && TInstant::Now() - start < TDuration::Seconds(10)) {
-                    Sleep(TDuration::Seconds(1));
-                    Cerr << "WAIT_PROPOSING..." << Endl;
-                }
-                AFL_VERIFY(CSTransferStatus->GetProposed());
-            }
-            if (RebootTablet) {
-                Kikimr.GetTestServer().GetRuntime()->Send(MakePipePeNodeCacheID(false), NActors::TActorId(), new TEvPipeCache::TEvForward(
-                    new TEvents::TEvPoisonPill(), destination, false));
-            }
-            {
-                const TInstant start = TInstant::Now();
-                while (!CSTransferStatus->GetConfirmed() && TInstant::Now() - start < TDuration::Seconds(10)) {
-                    Kikimr.GetTestServer().GetRuntime()->Send(MakePipePeNodeCacheID(false), NActors::TActorId(), new TEvPipeCache::TEvForward(
-                        new NOlap::NDataSharing::NEvents::TEvConfirmFromInitiator(sessionId), destination, false));
-                    Sleep(TDuration::Seconds(1));
-                    Cerr << "WAIT_CONFIRMED..." << Endl;
-                }
-                AFL_VERIFY(CSTransferStatus->GetConfirmed());
-            }
-            if (RebootTablet) {
-                Kikimr.GetTestServer().GetRuntime()->Send(MakePipePeNodeCacheID(false), NActors::TActorId(), new TEvPipeCache::TEvForward(
-                    new TEvents::TEvPoisonPill(), destination, false));
-                for (auto&& i : sources) {
-                    Kikimr.GetTestServer().GetRuntime()->Send(MakePipePeNodeCacheID(false), NActors::TActorId(), new TEvPipeCache::TEvForward(
-                        new TEvents::TEvPoisonPill(), i, false));
-                }
-            }
-            {
-                const TInstant start = TInstant::Now();
-                while (!CSTransferStatus->GetFinished() && TInstant::Now() - start < TDuration::Seconds(10)) {
-                    Sleep(TDuration::Seconds(1));
-                    Cerr << "WAIT_FINISHED..." << Endl;
-                }
-                AFL_VERIFY(CSTransferStatus->GetFinished());
-            }
-            CSTransferStatus->Reset();
-            AFL_VERIFY(!Controller->IsTrivialLinks());
-        }
-    };
-
-    Y_UNIT_TEST(BlobsSharingSplit1_1) {
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        TSharingDataTestCase tester(4, kikimr);
-        tester.AddRecords(800000);
-        Sleep(TDuration::Seconds(1));
-        tester.Execute(0, {1}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-    }
-
-    Y_UNIT_TEST(BlobsSharingSplit1_1_clean) {
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        TSharingDataTestCase tester(2, kikimr);
-        tester.AddRecords(80000);
-        CompareYson(tester.GetHelper().GetQueryResult("SELECT COUNT(*) FROM `/Root/olapStore12/olapTable`"), R"([[80000u;]])");
-        Sleep(TDuration::Seconds(1));
-        tester.Execute(0, {1}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-        CompareYson(tester.GetHelper().GetQueryResult("SELECT COUNT(*) FROM `/Root/olapStore12/olapTable`"), R"([[119928u;]])");
-        tester.AddRecords(80000, 0.8);
-        tester.WaitNormalization();
-        CompareYson(tester.GetHelper().GetQueryResult("SELECT COUNT(*) FROM `/Root/olapStore12/olapTable`"), R"([[183928u;]])");
-    }
-
-    Y_UNIT_TEST(BlobsSharingSplit1_1_clean_with_restarts) {
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        TSharingDataTestCase tester(2, kikimr);
-        tester.SetRebootTablet(true);
-        tester.AddRecords(80000);
-        CompareYson(tester.GetHelper().GetQueryResult("SELECT COUNT(*) FROM `/Root/olapStore12/olapTable`"), R"([[80000u;]])");
-        Sleep(TDuration::Seconds(1));
-        tester.Execute(0, {1}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-        CompareYson(tester.GetHelper().GetQueryResult("SELECT COUNT(*) FROM `/Root/olapStore12/olapTable`"), R"([[119928u;]])");
-        tester.AddRecords(80000, 0.8);
-        tester.WaitNormalization();
-        CompareYson(tester.GetHelper().GetQueryResult("SELECT COUNT(*) FROM `/Root/olapStore12/olapTable`"), R"([[183928u;]])");
-    }
-
-    Y_UNIT_TEST(BlobsSharingSplit3_1) {
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        TSharingDataTestCase tester(4, kikimr);
-        tester.AddRecords(800000);
-        Sleep(TDuration::Seconds(1));
-        tester.Execute(0, {1, 2, 3}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-    }
-
-    Y_UNIT_TEST(BlobsSharingSplit1_3_1) {
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        TSharingDataTestCase tester(4, kikimr);
-        tester.AddRecords(800000);
-        Sleep(TDuration::Seconds(1));
-        tester.Execute(1, {0}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-        tester.Execute(2, {0}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-        tester.Execute(3, {0}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-        tester.Execute(0, {1, 2, 3}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-    }
-
-    Y_UNIT_TEST(BlobsSharingSplit1_3_2_1_clean) {
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        TSharingDataTestCase tester(4, kikimr);
-        tester.AddRecords(800000);
-        Sleep(TDuration::Seconds(1));
-        tester.Execute(1, {0}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-        tester.Execute(2, {0}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-        tester.Execute(3, {0}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-        tester.AddRecords(800000, 0.9);
-        Sleep(TDuration::Seconds(1));
-        tester.Execute(3, {2}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-        tester.Execute(0, {1, 2}, false, NOlap::TSnapshot(TInstant::Now().MilliSeconds(), 1232123), {0});
-        tester.WaitNormalization();
-    }
-
     Y_UNIT_TEST(SelectLimit1ManyShards) {
         TPortManager tp;
         ui16 mbusport = tp.GetPort(2134);
@@ -4555,7 +1405,7 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
 
         runtime->SetObserverFunc(captureEvents);
         auto streamSender = runtime->AllocateEdgeActor();
-        SendRequest(*runtime, streamSender, MakeStreamRequest(streamSender, "SELECT * FROM `/Root/selectStore/selectTable` LIMIT 1;", false));
+        NDataShard::NKqpHelpers::SendRequest(*runtime, streamSender, NDataShard::NKqpHelpers::MakeStreamRequest(streamSender, "SELECT * FROM `/Root/selectStore/selectTable` LIMIT 1;", false));
         auto ev = runtime->GrabEdgeEventRethrow<NKqp::TEvKqp::TEvQueryResponse>(streamSender);
         UNIT_ASSERT_VALUES_EQUAL(result, 1);
     }
@@ -4621,7 +1471,7 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
 
         runtime->SetObserverFunc(captureEvents);
         auto streamSender = runtime->AllocateEdgeActor();
-        SendRequest(*runtime, streamSender, MakeStreamRequest(streamSender, "SELECT COUNT(*) FROM `/Root/largeOlapStore/largeOlapTable`;", false));
+        NDataShard::NKqpHelpers::SendRequest(*runtime, streamSender, NDataShard::NKqpHelpers::MakeStreamRequest(streamSender, "SELECT COUNT(*) FROM `/Root/largeOlapStore/largeOlapTable`;", false));
         runtime->GrabEdgeEventRethrow<NKqp::TEvKqp::TEvQueryResponse>(streamSender);
         UNIT_ASSERT_VALUES_EQUAL(result, insertRows);
     }
@@ -4684,7 +1534,7 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
 
         runtime->SetObserverFunc(captureEvents);
         auto streamSender = runtime->AllocateEdgeActor();
-        SendRequest(*runtime, streamSender, MakeStreamRequest(streamSender, "SELECT * FROM `/Root/largeOlapStore/largeOlapTable` where resource_id = Utf8(\"notfound\");", false));
+        NDataShard::NKqpHelpers::SendRequest(*runtime, streamSender, NDataShard::NKqpHelpers::MakeStreamRequest(streamSender, "SELECT * FROM `/Root/largeOlapStore/largeOlapTable` where resource_id = Utf8(\"notfound\");", false));
         auto ev = runtime->GrabEdgeEventRethrow<NKqp::TEvKqp::TEvQueryResponse>(streamSender);
         UNIT_ASSERT(hasResult);
     }
@@ -4781,382 +1631,9 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
 
         runtime->SetObserverFunc(captureEvents);
         auto streamSender = runtime->AllocateEdgeActor();
-        SendRequest(*runtime, streamSender, MakeStreamRequest(streamSender, "SELECT COUNT(*) FROM `/Root/largeOlapStore/largeOlapTable`;", false));
+        NDataShard::NKqpHelpers::SendRequest(*runtime, streamSender, NDataShard::NKqpHelpers::MakeStreamRequest(streamSender, "SELECT COUNT(*) FROM `/Root/largeOlapStore/largeOlapTable`;", false));
         auto ev = runtime->GrabEdgeEventRethrow<NKqp::TEvKqp::TEvQueryResponse>(streamSender);
         UNIT_ASSERT_VALUES_EQUAL(result, insertRows);
-    }
-
-    Y_UNIT_TEST(StatsSysViewColumns) {
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        TKikimrRunner kikimr(settings);
-
-        TLocalHelper(kikimr.GetTestServer()).CreateTestOlapTable();
-        for (ui64 i = 0; i < 10; ++i) {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 0, 1000000 + i*10000, 2000);
-        }
-
-        auto tableClient = kikimr.GetTableClient();
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT TabletId, PathId, Kind
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                ORDER BY PathId, Kind, TabletId
-                LIMIT 4;
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 4);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("PathId")), 3ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[0].at("Kind")), "INSERTED");
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[3].at("PathId")), 3ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[3].at("Kind")), "INSERTED");
-        }
-        {
-            auto selectQuery = TString(R"(
-                SELECT SUM(BlobRangeSize) as Bytes, SUM(Rows) as Rows, PathId, TabletId
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                GROUP BY PathId, TabletId
-                ORDER BY Bytes
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 3);
-            UNIT_ASSERT_LE(GetUint64(rows[0].at("Bytes")), GetUint64(rows[1].at("Bytes")));
-        }
-        {
-            auto selectQuery = TString(R"(
-                SELECT Sum(Rows) as Rows, Kind, Sum(RawBytes) as RawBytes, Sum(Rows) as Rows2, Sum(Rows) as Rows3, PathId
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                GROUP BY Kind, PathId
-                ORDER BY PathId, Kind, Rows3
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 2);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("Rows2")), GetUint64(rows[0].at("Rows3")));
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[1].at("Rows")), GetUint64(rows[1].at("Rows3")));
-        }
-    }
-
-    Y_UNIT_TEST(StatsSysViewRanges) {
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-        csController->SetCompactionControl(NYDBTest::EOptimizerCompactionWeightControl::Disable);
-        Tests::NCommon::TLoggerInit(kikimr).Initialize();
-
-        TLocalHelper(kikimr).CreateTestOlapTable("olapTable_1");
-        TLocalHelper(kikimr).CreateTestOlapTable("olapTable_2");
-        TLocalHelper(kikimr).CreateTestOlapTable("olapTable_3");
-
-        for (ui64 i = 0; i < 10; ++i) {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable_1", 0, 1000000 + i*10000, 2000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable_2", 0, 1000000 + i*10000, 3000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable_3", 0, 1000000 + i*10000, 5000);
-        }
-
-        auto tableClient = kikimr.GetTableClient();
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT *
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT PathId, Kind, TabletId
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                WHERE
-                    PathId == UInt64("3") AND Activity = true
-                GROUP BY TabletId, PathId, Kind
-                ORDER BY TabletId, Kind
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 3);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("PathId")), 3ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[0].at("Kind")), "INSERTED");
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[1].at("PathId")), 3ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[2].at("Kind")), "INSERTED");
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[2].at("PathId")), 3ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[1].at("Kind")), "INSERTED");
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT PathId, Kind, TabletId
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                GROUP BY PathId, Kind, TabletId
-                ORDER BY PathId DESC, Kind DESC, TabletId DESC
-                ;
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            ui32 numExpected = 3*3;
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), numExpected);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("PathId")), 5ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[0].at("Kind")), "INSERTED");
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[numExpected-1].at("PathId")), 3ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[numExpected-1].at("Kind")), "INSERTED");
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT PathId, Kind, TabletId
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                WHERE
-                    PathId > UInt64("0") AND PathId < UInt32("4")
-                    OR PathId > UInt64("4") AND PathId <= UInt64("5")
-                GROUP BY PathId, Kind, TabletId
-                ORDER BY
-                    PathId DESC, Kind DESC, TabletId DESC
-                ;
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            ui32 numExpected = 2*3;
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), numExpected);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("PathId")), 5ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[0].at("Kind")), "INSERTED");
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[numExpected-1].at("PathId")), 3ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUtf8(rows[numExpected-1].at("Kind")), "INSERTED");
-        }
-    }
-
-    Y_UNIT_TEST(StatsSysViewFilter) {
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-
-        TLocalHelper(kikimr.GetTestServer()).CreateTestOlapTable();
-        for (ui64 i = 0; i < 10; ++i) {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable", 0, 1000000 + i*10000, 2000);
-        }
-
-        auto tableClient = kikimr.GetTableClient();
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT PathId, Kind, TabletId, Sum(BlobRangeSize) as Bytes
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                GROUP BY PathId, Kind, TabletId
-                ORDER BY PathId, Kind, TabletId;
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            UNIT_ASSERT_GE(rows.size(), 3);
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT PathId, Kind, TabletId, Sum(BlobRangeSize) as Bytes
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                GROUP BY PathId, Kind, TabletId
-                ORDER BY PathId, Kind, TabletId;
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            UNIT_ASSERT_GE(rows.size(), 3);
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT *
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                WHERE Kind == 'EVICTED'
-                ORDER BY PathId, Kind, TabletId;
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            UNIT_ASSERT_GE(rows.size(), 0);
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT PathId, Kind, TabletId
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                WHERE Kind IN ('SPLIT_COMPACTED', 'INACTIVE', 'EVICTED')
-                GROUP BY PathId, Kind, TabletId
-                ORDER BY PathId, Kind, TabletId;
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-
-            UNIT_ASSERT_GE(rows.size(), 3);
-        }
-    }
-
-    Y_UNIT_TEST(StatsSysViewAggregation) {
-        auto settings = TKikimrSettings().SetWithSampleTables(false);
-        TKikimrRunner kikimr(settings);
-        auto csController = NYDBTest::TControllers::RegisterCSControllerGuard<NYDBTest::NColumnShard::TController>();
-
-        TLocalHelper(kikimr.GetTestServer()).CreateTestOlapTable("olapTable_1");
-        TLocalHelper(kikimr.GetTestServer()).CreateTestOlapTable("olapTable_2");
-        TLocalHelper(kikimr.GetTestServer()).CreateTestOlapTable("olapTable_3");
-
-        for (ui64 i = 0; i < 100; ++i) {
-            WriteTestData(kikimr, "/Root/olapStore/olapTable_1", 0, 1000000 + i*10000, 1000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable_2", 0, 1000000 + i*10000, 2000);
-            WriteTestData(kikimr, "/Root/olapStore/olapTable_3", 0, 1000000 + i*10000, 3000);
-        }
-
-        Tests::NCommon::TLoggerInit(kikimr).Initialize();
-
-        auto tableClient = kikimr.GetTableClient();
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT
-                    SUM(Rows) as rows,
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                WHERE
-                    Kind != 'INACTIVE'
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 1ull);
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT
-                    PathId,
-                    SUM(Rows) as rows,
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                WHERE
-                    Kind != 'INACTIVE'
-                GROUP BY
-                    PathId
-                ORDER BY
-                    PathId
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 3ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("PathId")), 3);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[1].at("PathId")), 4);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[2].at("PathId")), 5);
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT
-                    PathId,
-                    SUM(Rows) as rows,
-                    SUM(BlobRangeSize) as bytes,
-                    SUM(RawBytes) as bytes_raw
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                WHERE
-                    Kind IN ('INSERTED', 'SPLIT_COMPACTED', 'COMPACTED')
-                GROUP BY PathId
-                ORDER BY rows DESC
-                LIMIT 10
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 3ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("PathId")), 5);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[1].at("PathId")), 4);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[2].at("PathId")), 3);
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT
-                    PathId,
-                    SUM(Rows) as rows,
-                    SUM(BlobRangeSize) as bytes,
-                    SUM(RawBytes) as bytes_raw
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                WHERE
-                    PathId == UInt64("3") AND Kind IN ('INSERTED', 'SPLIT_COMPACTED', 'COMPACTED')
-                GROUP BY PathId
-                ORDER BY rows DESC
-                LIMIT 10
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 1ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("PathId")), 3);
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT
-                    PathId,
-                    SUM(Rows) as rows,
-                    SUM(BlobRangeSize) as bytes,
-                    SUM(RawBytes) as bytes_raw
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                WHERE
-                    PathId >= UInt64("4") AND Kind IN ('INSERTED', 'SPLIT_COMPACTED', 'COMPACTED')
-                GROUP BY PathId
-                ORDER BY rows DESC
-                LIMIT 10
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 2ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("PathId")), 5);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[1].at("PathId")), 4);
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT PathId, TabletId, Kind
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                GROUP BY PathId, TabletId, Kind
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-            // 3 Tables with 3 Shards each and 2 KindId-s of stats
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 3 * 3 * 2);
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT
-                    count(distinct(PathId)),
-                    count(distinct(Kind)),
-                    count(distinct(TabletId))
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("column0")), 3ull);
-            UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[0].at("column1")), 2);
-            UNIT_ASSERT_GE(GetUint64(rows[0].at("column2")), 3ull);
-        }
-
-        {
-            auto selectQuery = TString(R"(
-                SELECT PathId, count(*), sum(Rows), sum(BlobRangeSize), sum(RawBytes)
-                FROM `/Root/olapStore/.sys/store_primary_index_stats`
-                GROUP BY PathId
-                ORDER BY PathId
-            )");
-
-            auto rows = ExecuteScanQuery(tableClient, selectQuery);
-            UNIT_ASSERT_VALUES_EQUAL(rows.size(), 3ull);
-            for (ui64 pathId = 3, row = 0; pathId <= 5; ++pathId, ++row) {
-                UNIT_ASSERT_VALUES_EQUAL(GetUint64(rows[row].at("PathId")), pathId);
-            }
-        }
     }
 
     Y_UNIT_TEST(PredicatePushdownWithParameters) {
@@ -5733,211 +2210,6 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
 
         b << "------------------------------------------------" << Endl;
         UNIT_ASSERT_C(falsePositive.empty() && falseNegative.empty(), b);
-    }
-
-    Y_UNIT_TEST(NoErrorOnLegacyPragma) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                PRAGMA Kikimr.KqpPushOlapProcess = "false";
-                SELECT id, resource_id FROM `/Root/tableWithNulls`
-                WHERE
-                    level = 5;
-            )")
-            .SetExpectedReply("[[5;#]]")
-            .AddExpectedPlanOptions("KqpOlapFilter");
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(BlocksRead) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                PRAGMA UseBlocks;
-                PRAGMA Kikimr.OptEnableOlapPushdown = "false";
-
-                SELECT
-                    id, resource_id
-                FROM `/Root/tableWithNulls`
-                WHERE
-                    level = 5;
-            )")
-            .SetExpectedReply("[[5;#]]");
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Blocks_NoAggPushdown) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                PRAGMA UseBlocks;
-                SELECT
-                    COUNT(DISTINCT id)
-                FROM `/Root/tableWithNulls`;
-            )")
-            .SetExpectedReply("[[10u]]");
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Json_GetValue) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT id, JSON_VALUE(jsonval, "$.col1"), JSON_VALUE(jsondoc, "$.col1") FROM `/Root/tableWithNulls`
-                WHERE JSON_VALUE(jsonval, "$.col1") = "val1" AND id = 1;
-            )")
-#if SSA_RUNTIME_VERSION >= 3U
-            .AddExpectedPlanOptions("KqpOlapJsonValue")
-#else
-            .AddExpectedPlanOptions("Udf")
-#endif
-            .SetExpectedReply(R"([[1;["val1"];#]])");
-
-        TestTableWithNulls({testCase});
-    }
-
-    Y_UNIT_TEST(Json_GetValue_Minus) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT id, JSON_VALUE(jsonval, "$.'col-abc'"), JSON_VALUE(jsondoc, "$.'col-abc'") FROM `/Root/tableWithNulls`
-                WHERE JSON_VALUE(jsonval, "$.'col-abc'") = "val-abc" AND id = 1;
-            )")
-#if SSA_RUNTIME_VERSION >= 3U
-            .AddExpectedPlanOptions("KqpOlapJsonValue")
-#else
-            .AddExpectedPlanOptions("Udf")
-#endif
-            .SetExpectedReply(R"([[1;["val-abc"];#]])");
-
-        TestTableWithNulls({testCase});
-    }
-
-    Y_UNIT_TEST(Json_GetValue_ToString) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT id, JSON_VALUE(jsonval, "$.col1" RETURNING String), JSON_VALUE(jsondoc, "$.col1") FROM `/Root/tableWithNulls`
-                WHERE JSON_VALUE(jsonval, "$.col1" RETURNING String) = "val1" AND id = 1;
-            )")
-#if SSA_RUNTIME_VERSION >= 3U
-            .AddExpectedPlanOptions("KqpOlapJsonValue")
-#else
-            .AddExpectedPlanOptions("Udf")
-#endif
-            .SetExpectedReply(R"([[1;["val1"];#]])");
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Json_GetValue_ToInt) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT id, JSON_VALUE(jsonval, "$.obj.obj_col2_int" RETURNING Int), JSON_VALUE(jsondoc, "$.obj.obj_col2_int" RETURNING Int) FROM `/Root/tableWithNulls`
-                WHERE JSON_VALUE(jsonval, "$.obj.obj_col2_int" RETURNING Int) = 16 AND id = 1;
-            )")
-#if SSA_RUNTIME_VERSION >= 3U
-            .AddExpectedPlanOptions("KqpOlapJsonValue")
-#else
-            .AddExpectedPlanOptions("Udf")
-#endif
-            .SetExpectedReply(R"([[1;[16];#]])");
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(JsonDoc_GetValue) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT id, JSON_VALUE(jsonval, "$.col1"), JSON_VALUE(jsondoc, "$.col1") FROM `/Root/tableWithNulls`
-                WHERE JSON_VALUE(jsondoc, "$.col1") = "val1" AND id = 6;
-            )")
-#if SSA_RUNTIME_VERSION >= 3U
-            .AddExpectedPlanOptions("KqpOlapJsonValue")
-#else
-            .AddExpectedPlanOptions("Udf")
-#endif
-            .SetExpectedReply(R"([[6;#;["val1"]]])");
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(JsonDoc_GetValue_ToString) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT id, JSON_VALUE(jsonval, "$.col1"), JSON_VALUE(jsondoc, "$.col1" RETURNING String) FROM `/Root/tableWithNulls`
-                WHERE JSON_VALUE(jsondoc, "$.col1" RETURNING String) = "val1" AND id = 6;
-            )")
-#if SSA_RUNTIME_VERSION >= 3U
-            .AddExpectedPlanOptions("KqpOlapJsonValue")
-#else
-            .AddExpectedPlanOptions("Udf")
-#endif
-            .SetExpectedReply(R"([[6;#;["val1"]]])");
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(JsonDoc_GetValue_ToInt) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT id, JSON_VALUE(jsonval, "$.obj.obj_col2_int"), JSON_VALUE(jsondoc, "$.obj.obj_col2_int" RETURNING Int) FROM `/Root/tableWithNulls`
-                WHERE JSON_VALUE(jsondoc, "$.obj.obj_col2_int" RETURNING Int) = 16 AND id = 6;
-            )")
-#if SSA_RUNTIME_VERSION >= 3U
-            .AddExpectedPlanOptions("KqpOlapJsonValue")
-#else
-            .AddExpectedPlanOptions("Udf")
-#endif
-            .SetExpectedReply(R"([[6;#;[16]]])");
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Json_Exists) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT id, JSON_EXISTS(jsonval, "$.col1"), JSON_EXISTS(jsondoc, "$.col1") FROM `/Root/tableWithNulls`
-                WHERE
-                    JSON_EXISTS(jsonval, "$.col1") AND level = 1;
-            )")
-#if SSA_RUNTIME_VERSION >= 3U
-            .AddExpectedPlanOptions("KqpOlapJsonExists")
-#else
-            .AddExpectedPlanOptions("Udf")
-#endif
-            .SetExpectedReply(R"([[1;[%true];#]])");
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(JsonDoc_Exists) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT id, JSON_EXISTS(jsonval, "$.col1"), JSON_EXISTS(jsondoc, "$.col1") FROM `/Root/tableWithNulls`
-                WHERE
-                    JSON_EXISTS(jsondoc, "$.col1") AND id = 6;
-            )")
-#if SSA_RUNTIME_VERSION >= 3U
-            .AddExpectedPlanOptions("KqpOlapJsonExists")
-#else
-            .AddExpectedPlanOptions("Udf")
-#endif
-            .SetExpectedReply(R"([[6;#;[%true]]])");
-
-        TestTableWithNulls({ testCase });
-    }
-
-    Y_UNIT_TEST(Json_Query) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT id, JSON_QUERY(jsonval, "$.col1" WITH UNCONDITIONAL WRAPPER),
-                    JSON_QUERY(jsondoc, "$.col1" WITH UNCONDITIONAL WRAPPER)
-                FROM `/Root/tableWithNulls`
-                WHERE
-                    level = 1;
-            )")
-            .AddExpectedPlanOptions("Udf")
-            .SetExpectedReply(R"([[1;["[\"val1\"]"];#]])");
-
-        TestTableWithNulls({ testCase });
     }
 
     Y_UNIT_TEST(Olap_InsertFailsOnDataQuery) {
@@ -6739,53 +3011,6 @@ Y_UNIT_TEST_SUITE(KqpOlap) {
         }
     }
 
-    Y_UNIT_TEST(BlockGenericWithDistinct) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    COUNT(DISTINCT id)
-                FROM `/Root/tableWithNulls`
-                WHERE level = 5 AND Cast(id AS String) = "5";
-            )")
-            .AddExpectedPlanOptions("KqpBlockReadOlapTableRanges")
-            .AddExpectedPlanOptions("WideFromBlocks")
-            .SetExpectedReply("[[1u]]");
-        TestTableWithNulls({ testCase }, /* generic */ true);
-    }
-
-    Y_UNIT_TEST(BlockGenericSimpleAggregation) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    level, COUNT(*), SUM(id)
-                FROM `/Root/tableWithNulls`
-                WHERE level = 5
-                GROUP BY level
-                ORDER BY level;
-            )")
-            .AddExpectedPlanOptions("KqpBlockReadOlapTableRanges")
-            .AddExpectedPlanOptions("WideFromBlocks")
-            .SetExpectedReply(R"([[[5];1u;5]])");
-
-        TestTableWithNulls({ testCase }, /* generic */ true);
-    }
-
-    Y_UNIT_TEST(BlockGenericSelectAll) {
-        TAggregationTestCase testCase;
-        testCase.SetQuery(R"(
-                SELECT
-                    id, resource_id, level
-                FROM `/Root/tableWithNulls`
-                WHERE level != 5 OR level IS NULL
-                ORDER BY id, resource_id, level;
-            )")
-            .AddExpectedPlanOptions("KqpBlockReadOlapTableRanges")
-            .AddExpectedPlanOptions("WideFromBlocks")
-            .SetExpectedReply(R"([[1;#;[1]];[2;#;[2]];[3;#;[3]];[4;#;[4]];[6;["6"];#];[7;["7"];#];[8;["8"];#];[9;["9"];#];[10;["10"];#]])");
-
-        TestTableWithNulls({ testCase }, /* generic */ true);
-    }
 }
 
-} // namespace NKqp
-} // namespace NKikimr
+}
