+++ ydb/core/kqp/ut/olap/ya.make
@@ -15,6 +15,12 @@ ENDIF()
 SRCS(
     kqp_olap_stats_ut.cpp
     GLOBAL kqp_olap_ut.cpp
+    sys_view_ut.cpp
+    indexes_ut.cpp
+    GLOBAL blobs_sharing_ut.cpp
+    statistics_ut.cpp
+    clickbench_ut.cpp
+    aggregations_ut.cpp
 )
 
 PEERDIR(
@@ -22,6 +28,7 @@ PEERDIR(
     ydb/core/kqp/ut/common
     ydb/library/yql/sql/pg_dummy
     ydb/core/tx/columnshard/hooks/testing
+    ydb/core/kqp/ut/olap/helpers
     ydb/core/tx/datashard/ut_common
 )
